% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISADataFrame.R
\name{add_metadata}
\alias{add_metadata}
\title{Adds new metadata to the metadata attribute.}
\usage{
add_metadata(x, meta)
}
\arguments{
\item{x}{The ISADataFrame object}

\item{meta}{A character vector containing the names of the columns to
register as metadata}
}
\value{
An ISADataFrame object
}
\description{
The metadata has to be present in the data frame in order to be added.
}
\examples{
example_matrix_ann <- system.file("extdata", "ex_annotated_ISMatrix.tsv.xz",
    package = "ISAnalytics"
)
isadf_ann <- import_single_Vispa2Matrix(example_matrix_ann)
isadf_ann <- add_metadata(isadf_ann, c("GeneName"))
}
