% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported-vars.R
\name{annotation_IS_vars}
\alias{annotation_IS_vars}
\title{Names of the annotation variables for an integration matrix.}
\usage{
annotation_IS_vars()
}
\value{
A character vector
}
\description{
Contains the names of the columns that are present if the integration matrix
is annotated.
}
\examples{
annotation_IS_vars()
}
