% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{association_file}
\alias{association_file}
\title{Example of association file.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 53 rows and 83 columns.
}
\usage{
data("association_file")
}
\description{
This file is a simple example of association file. Use it as
reference to properly fill out yours.
To generate an empty association file to fill see the
\code{generate_blank_association_file()} function.
}
\details{
The data was obtained manually by simulating real research
data.
}
\seealso{
\code{\link{generate_blank_association_file}}
}
\keyword{datasets}
