% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{association_file_columns}
\alias{association_file_columns}
\title{Names of the columns in the association file.}
\usage{
association_file_columns(include_types = FALSE)
}
\arguments{
\item{include_types}{If set to \code{TRUE} returns both the names and the types
associated, otherwise returns only a character vector of names}
}
\value{
A character vector or a data frame
}
\description{
All the names of the columns present in the association file.
}
\examples{
# Names only
association_file_columns()

# Names and types
association_file_columns(TRUE)
}
