% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier-filtering.R
\name{available_outlier_tests}
\alias{available_outlier_tests}
\title{A character vector containing all the names of the currently supported
outliers tests that can be called in the function \link{outlier_filter}.}
\usage{
available_outlier_tests()
}
\value{
A character vector
}
\description{
A character vector containing all the names of the currently supported
outliers tests that can be called in the function \link{outlier_filter}.
}
\examples{
available_outlier_tests()
}
\seealso{
Other Outlier tests: 
\code{\link{outliers_by_pool_fragments}()}
}
\concept{Outlier tests}
