% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-size-estimate.R
\name{blood_lineages_default}
\alias{blood_lineages_default}
\title{Default blood lineages info}
\usage{
blood_lineages_default()
}
\value{
A data frame
}
\description{
A default table with info relative to different blood lineages associated
with cell markers that can be supplied as a parameter to
\code{\link{HSC_population_size_estimate}}
}
\examples{
blood_lineages_default()
}
