% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISADataFrame.R
\name{check_metadata}
\alias{check_metadata}
\title{Internal helper function for checking metadata fields.}
\usage{
check_metadata(x)
}
\arguments{
\item{x}{an ISADataFrame object}
}
\value{
FALSE if some metadata are not found in the data frame, TRUE otherwise
}
\description{
Checks if the specified metadata are present in the data frame.
}
\examples{
\dontrun{
isadf <- new_ISADataFrame(list, meta = c("meta1"))
check_result <- check_metadata(isadf)
}
}
