% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{clinical_relevant_suspicious_genes}
\alias{clinical_relevant_suspicious_genes}
\title{Clinical relevant suspicious genes (for mouse and human).}
\usage{
clinical_relevant_suspicious_genes()
}
\value{
A data frame
}
\description{
Clinical relevant suspicious genes (for mouse and human).
}
\examples{
clinical_relevant_suspicious_genes()
}
\seealso{
Other Plotting function helpers: 
\code{\link{known_clinical_oncogenes}()}
}
\concept{Plotting function helpers}
