% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collision-removal.R
\name{date_columns_coll}
\alias{date_columns_coll}
\title{Possible choices for \code{date_col} parameter.}
\usage{
date_columns_coll()
}
\value{
A character vector of column names
}
\description{
Possible choices for \code{date_col} parameter.
}
\examples{
dates <- date_columns_coll()
}
\seealso{
\code{\link{remove_collisions}}
}
\concept{Collision removal helpers}
