% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\name{date_formats}
\alias{date_formats}
\title{Possible choices for the \code{dates_format} parameter in
\code{import_association_file},
\code{import_parallel_vispa2Matrices_interactive} and
\code{import_parallel_vispa2Matrices_auto}.}
\usage{
date_formats()
}
\value{
A character vector
}
\description{
All options correspond to \code{lubridate} functions:
\itemize{
\item ymd: year, month, date
\item ydm: year, day, month
\item mdy: month, day, year
\item myd: month, year, day
\item dmy: day, month, year - default value
\item dym: day, year, month
\item yq: year quantile
}
}
\details{
\strong{NOTE: use the same date format across the association file.}
}
\examples{
date_formats()
}
\seealso{
\code{\link{import_association_file}},
\code{\link{import_parallel_Vispa2Matrices_auto}}
}
