% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\name{default_af_transform}
\alias{default_af_transform}
\title{Default transformations to apply to association file columns.}
\usage{
default_af_transform(convert_tp)
}
\arguments{
\item{convert_tp}{The value of the argument \code{convert_tp} in the call
to \code{import_association_file()}}
}
\value{
A named list of lambdas
}
\description{
A list of default transformations to apply to the association file columns
after importing it via \code{import_association_file()}
}
\examples{
default_af_transform(TRUE)
}
\seealso{
Other Import functions helpers: 
\code{\link{annotation_issues}()},
\code{\link{date_formats}()},
\code{\link{default_iss_file_prefixes}()},
\code{\link{matching_options}()},
\code{\link{quantification_types}()}
}
\concept{Import functions helpers}
