% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\name{default_iss_file_prefixes}
\alias{default_iss_file_prefixes}
\title{Default regex prefixes for Vispa2 stats files.}
\usage{
default_iss_file_prefixes()
}
\value{
A character vector of regexes
}
\description{
Note that each element is a regular expression.
}
\examples{
default_iss_file_prefixes()
}
