% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate-functions.R
\name{default_meta_agg}
\alias{default_meta_agg}
\title{Default metadata aggregation function table}
\usage{
default_meta_agg()
}
\value{
A data frame
}
\description{
A default columns-function specifications for \link{aggregate_metadata}
}
\details{
This data frame contains four columns:
\itemize{
\item \code{Column}: holds the name of the column in the association file that
should be processed
\item \code{Function}: contains either the name of a function (e.g. mean)
or a purrr-style lambda (e.g. \code{~ mean(.x, na.rm = TRUE)}). This function
will be applied to the corresponding column specified in \code{Column}
\item \code{Args}: optional additional arguments to pass to the corresponding
function. This is relevant ONLY if the corresponding \code{Function} is a
simple function and not a purrr-style lambda.
\item \code{Output_colname}: a \code{glue} specification that will be used to determine
a unique output column name. See \link[glue]{glue} for more details.
}
}
\examples{
default_meta_agg()
}
\seealso{
Other Data cleaning and pre-processing: 
\code{\link{aggregate_metadata}()},
\code{\link{aggregate_values_by_key}()},
\code{\link{compute_near_integrations}()},
\code{\link{outlier_filter}()},
\code{\link{outliers_by_pool_fragments}()},
\code{\link{purity_filter}()},
\code{\link{realign_after_collisions}()},
\code{\link{remove_collisions}()},
\code{\link{threshold_filter}()}
}
\concept{Data cleaning and pre-processing}
