% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalibration-functions.R
\name{default_rec_agg_lambdas}
\alias{default_rec_agg_lambdas}
\title{Defaults for column aggregations in \code{compute_near_integrations()}.}
\usage{
default_rec_agg_lambdas()
}
\value{
A named list of lambdas
}
\description{
Defaults for column aggregations in \code{compute_near_integrations()}.
}
\examples{
default_rec_agg_lambdas()
}
