% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-utilities.R
\name{default_report_path}
\alias{default_report_path}
\title{Default folder for saving ISAnalytics reports. Supplied as default
argument for several functions.}
\usage{
default_report_path()
}
\value{
A path
}
\description{
Default folder for saving ISAnalytics reports. Supplied as default
argument for several functions.
}
\examples{
default_report_path()
}
