% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-functions.R
\name{default_stats}
\alias{default_stats}
\title{A set of pre-defined functions for \code{sample_statistics}.}
\usage{
default_stats()
}
\value{
A named list of functions/purrr-style lambdas
}
\description{
A set of pre-defined functions for \code{sample_statistics}.
}
\examples{
default_stats()
}
\concept{Analysis functions helpers}
