% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.aggregate_lambda}
\alias{.aggregate_lambda}
\title{Internal function that performs aggregation on values with a lambda
operation.}
\usage{
.aggregate_lambda(x, af, key, lambda, group, args, namespace, envir)
}
\arguments{
\item{x}{The list of matrices to aggregate. If a single matrix has to be
supplied it must be enclosed in a list. For example \code{x = list(matrix)}.}

\item{af}{The association file}

\item{key}{A string or character vector to use as key}

\item{lambda}{The aggregating operation to apply to values. Must take as
input a numeric/integer vector and return a single value}

\item{group}{The additional variables to add to grouping}

\item{args}{Additional arguments passed on to lambda (named list)}

\item{namespace}{The namespace from which the lambda is exported}

\item{envir}{The environment in which symbols must be evaluated}
}
\value{
A list of tibbles with aggregated values
}
\description{
Internal function that performs aggregation on values with a lambda
operation.
}
\keyword{internal}
