% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.all_match}
\alias{.all_match}
\title{Helper function for checking if all of the elements of the list is true.}
\usage{
.all_match(...)
}
\arguments{
\item{...}{A list of logical values}
}
\value{
TRUE if all of the parameters is true
}
\description{
Helper function for checking if all of the elements of the list is true.
}
\keyword{internal}
