% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.any_match}
\alias{.any_match}
\title{Helper function for checking if any of the elements of the list is true.}
\usage{
.any_match(...)
}
\arguments{
\item{...}{A list of logical values}
}
\value{
TRUE if any of the parameters is true
}
\description{
Helper function for checking if any of the elements of the list is true.
}
