% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.check_af_correctness}
\alias{.check_af_correctness}
\title{Checks if the association file has the right format (standard headers).}
\usage{
.check_af_correctness(df)
}
\arguments{
\item{df}{The imported association file}
}
\value{
TRUE if the check passes, FALSE otherwise
}
\description{
Checks if the association file has the right format (standard headers).
}
