% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISADataFrame.R
\name{.check_atLeastOneExp}
\alias{.check_atLeastOneExp}
\title{Internal helper function for checking presence of experimental data columns.}
\usage{
.check_atLeastOneExp(x)
}
\arguments{
\item{x}{an ISADataFrame object}
}
\value{
FALSE if no experimental columns were detected, TRUE if at least
one was detected
}
\description{
Checks if there is at least one experimental data column
}
