% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.check_complAmpID}
\alias{.check_complAmpID}
\title{Internal helper function for checking \code{CompleteAmplifcationID}
column presence in x.}
\usage{
.check_complAmpID(x)
}
\arguments{
\item{x}{A data.frame object (or any extending class)}
}
\value{
FALSE if not found, TRUE otherwise
}
\description{
Checks if the column \code{CompleteAmplifcationID} is present in the data frame.
}
\keyword{internal}
