% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.check_file_system_alignment}
\alias{.check_file_system_alignment}
\title{Internal function to check alignment between association file and file
system starting from the root. The alignment is checked at folder level.}
\usage{
.check_file_system_alignment(df, root_folder)
}
\arguments{
\item{df}{The imported association file (data.frame or tibble)}

\item{root_folder}{Path to the root folder}
}
\value{
A data frame containing, for each ProjectID and
concatenatePoolIDSeqRun the
corresponding path on disk if found, NA otherwise.
}
\description{
Internal function to check alignment between association file and file
system starting from the root. The alignment is checked at folder level.
}
\keyword{internal}
