% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.check_mandatory_vars}
\alias{.check_mandatory_vars}
\title{Internal helper function for checking mandatory vars presence in x.}
\usage{
.check_mandatory_vars(x)
}
\arguments{
\item{x}{A data.frame object (or any extending class)}
}
\value{
FALSE if all or some elements are not found in the data frame, TRUE
otherwise
}
\description{
Checks if the elements of \code{mandatory_IS_vars} are present as column names
in the data frame.
}
\keyword{internal}
