% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.check_metadata}
\alias{.check_metadata}
\title{Internal helper function for checking metadata fields.}
\usage{
.check_metadata(x)
}
\arguments{
\item{x}{An ISADataFrame object}
}
\value{
FALSE if some metadata are not found in the data frame, TRUE
otherwise
}
\description{
Checks if the specified metadata are present in the data frame.
}
