% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.check_nonNumdata}
\alias{.check_nonNumdata}
\title{Internal helper function for checking presence of experimental data columns
and verifying they're numeric.}
\usage{
.check_nonNumdata(x)
}
\arguments{
\item{x}{an ISADataFrame object}
}
\value{
TRUE if at least one experimental data column of type numeric (or
integer) was detected, 'Warning' if one or more non-numeric
columns were detected which are not metadata, FALSE in all other cases
}
\description{
Internal helper function for checking presence of experimental data columns
and verifying they're numeric.
}
