% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.check_same_info}
\alias{.check_same_info}
\title{Checks if association file contains more information than the matrix.}
\usage{
.check_same_info(association_file, df)
}
\arguments{
\item{association_file}{The imported association file}

\item{df}{The sequence count matrix to examine}
}
\value{
A tibble containing ProjectID, PoolID and CompleteAmplificationID
only for additional info and only for projects already present in df
}
\description{
Used to notify the user that wants to know if for the projects contained in
the examined matrix there are additional CompleteAmplificationIDs contained
in the association file that weren't included in the integration matrix (for
example failed runs).
}
