% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.check_stats}
\alias{.check_stats}
\title{Checks if the stats file contains the minimal set of columns.}
\usage{
.check_stats(x)
}
\arguments{
\item{x}{The stats df}
}
\value{
TRUE or FALSE
}
\description{
Checks if the stats file contains the minimal set of columns.
}
\keyword{internal}
