% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.check_value_col}
\alias{.check_value_col}
\title{Internal helper function for checking \code{Value} column presence in x.}
\usage{
.check_value_col(x)
}
\arguments{
\item{x}{A data.frame object (or any extending class)}
}
\value{
FALSE if not found or contains non-numeric data, TRUE otherwise
}
\description{
Checks if the column \code{Value} is present in the data frame and also
checks if the column is numeric or integer.
}
\keyword{internal}
