% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.check_window_criteria}
\alias{.check_window_criteria}
\title{Wrapper function for criteria evaluation inside \code{window} function.}
\usage{
.check_window_criteria(criterias, values, subset)
}
\arguments{
\item{criterias}{The vector of criterias}

\item{values}{The numeric vector containing values}

\item{subset}{Integer vector with indexes of the values vector to pass to
individual functions.}
}
\value{
Values to drop
}
\description{
Wrapper function for criteria evaluation inside \code{window} function.
}
\keyword{internal}
