% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.checker_widget}
\alias{.checker_widget}
\title{Builds the html widget for the checker table.}
\usage{
.checker_widget(checker_df)
}
\arguments{
\item{checker_df}{Tibble obtained via \code{.check_file_system_alignment}}
}
\value{
An html widget
}
\description{
Builds the html widget for the checker table.
}
\keyword{internal}
