% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.choose_duplicates_files_interactive}
\alias{.choose_duplicates_files_interactive}
\title{Simple function to manage user input for duplicate file choices for each
quantification type in a single project/pool pair (use internally in
\code{.manage_anomalies_interactive}).}
\usage{
.choose_duplicates_files_interactive(q_types, dupl)
}
\arguments{
\item{q_types}{Vector of characters containing the unique quantification
types that are detected as duplicates}

\item{dupl}{The tibble containing quantification types and path to the files
found for a single project/pool pair}
}
\value{
An updated tibble containing files chosen for each type
}
\description{
Simple function to manage user input for duplicate file choices for each
quantification type in a single project/pool pair (use internally in
\code{.manage_anomalies_interactive}).
}
\keyword{internal}
