% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.coll_mapping}
\alias{.coll_mapping}
\title{Internal function to process collisions on multiple integrations.}
\usage{
.coll_mapping(x, date_col, reads_ratio, seqCount_col)
}
\arguments{
\item{x}{The nested collision data frame (or chunks for parallel execution)}

\item{date_col}{The date column to consider}

\item{reads_ratio}{The value of the ratio between sequence count values to
check}

\item{seqCount_col}{The name of the sequence count column (support
for multi quantification matrix)}
}
\value{
A list containing the updated collisions, a numeric value
representing the number of integrations removed and a numeric value
representing the number of integrations reassigned
}
\description{
Internal function to process collisions on multiple integrations.
}
\keyword{internal}
