% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.discriminate_by_date}
\alias{.discriminate_by_date}
\title{Internal for date discrimination in collision removal.}
\usage{
.discriminate_by_date(nest, date_col)
}
\arguments{
\item{nest}{The nested table associated with a single integration}

\item{date_col}{The name of the date column chosen for collision removal,
one in \code{date_columns_coll()}}
}
\value{
A named list with:
\itemize{
\item $data: a tibble, containing the data (unmodified or modified)
\item $check: a logical value indicating whether the analysis was successful or
not (and therefore there is the need to perform the next step)
}
}
\description{
It's the first of 4 steps in the algorithm for collision removal: it tries to
find a single sample who has an associated date which is earlier than any
other. If comparison is not possible the analysis fails and returns the
orginal input data.
}
\details{
NOTE: this function is meant to be used inside a mapping function
such as \code{purrr::pmap}. The function only works on data regarding a SINGLE
integration (triplet chr, integration_locus, strand).
}
\keyword{internal}
