% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.discriminate_by_replicate}
\alias{.discriminate_by_replicate}
\title{Internal for replicate discrimination in collision removal.}
\usage{
.discriminate_by_replicate(nest)
}
\arguments{
\item{nest}{The nested table associated with a single integration}
}
\value{
A named list with:
\itemize{
\item $data: a tibble, containing the data (unmodified or modified)
\item $check: a logical value indicating whether the analysis was successful or
not (and therefore there is the need to perform the next step)
}
}
\description{
It's the second of 4 steps in the algorithm for collision removal:
grouping by independent sample (ProjectID, SubjectID) it counts the number of
rows (replicates) found for each group, orders them from biggest to smallest
and, if a single group has more rows than any other group the integration is
assigned to that sample, otherwise the analysis fails and returns the
original input to be submitted to the next step.
}
\details{
NOTE: this function is meant to be used inside a mapping function
such as \code{purrr::pmap}. The function only works on data regarding a SINGLE
integration (triplet chr, integration_locus, strand).
}
