% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.discriminate_by_seqCount}
\alias{.discriminate_by_seqCount}
\title{Internal for sequence count discrimination in collision removal.}
\usage{
.discriminate_by_seqCount(nest, reads_ratio)
}
\arguments{
\item{nest}{The nested table associated with a single integration}

\item{reads_ratio}{The value of the ratio between sequence count values to
check}
}
\value{
A named list with:
\itemize{
\item $data: a tibble, containing the data (unmodified or modified)
\item $check: a logical value indicating whether the analysis was successful or
not (and therefore there is the need to perform the next step)
}
}
\description{
It's the third of 4 steps in the algorithm for collision removal:
grouping by independent sample (ProjectID, SubjectID), it sums the value of
the sequence count for each group, sorts the groups from highest to lowest
cumulative value and then checks the ratio between the first element and the
second: if the ratio is > \code{reads_ratio} the integration is assigned to
the first group, otherwise the analysis fails and returns the original input.
}
\details{
NOTE: this function is meant to be used inside a mapping function
such as \code{purrr::pmap}. The function only works on data regarding a SINGLE
integration (triplet chr, integration_locus, strand).
}
\keyword{internal}
