% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-widgets.R
\name{.files_found_widget}
\alias{.files_found_widget}
\title{Builds the html widget for the files_found table.}
\usage{
.files_found_widget(files_found)
}
\arguments{
\item{files_found}{Tibble obtained via \code{.lookup_matrices} or
\code{.lookup_matrices_auto}}
}
\value{
An html widget
}
\description{
Builds the html widget for the files_found table.
}
\keyword{internal}
