% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.files_imported_widget}
\alias{.files_imported_widget}
\title{Builds the html widget for the files_imported table.}
\usage{
.files_imported_widget(files_imported)
}
\arguments{
\item{files_imported}{Tibble obtained via \code{.parallel_import_merge}}
}
\value{
An html widget
}
\description{
Builds the html widget for the files_imported table.
}
