% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.files_to_import_widget}
\alias{.files_to_import_widget}
\title{Builds the html widget for the files_to_import table.}
\usage{
.files_to_import_widget(files_to_import)
}
\arguments{
\item{files_to_import}{Tibble obtained via
\code{.manage_anomalies_interactive} or
\code{.manage_anomalies_auto}}
}
\value{
An html widget
}
\description{
Builds the html widget for the files_to_import table.
}
\keyword{internal}
