% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.find_exp_cols}
\alias{.find_exp_cols}
\title{Finds experimental columns in an integration matrix.}
\usage{
.find_exp_cols(x)
}
\arguments{
\item{x}{A data.frame}
}
\value{
A character vector of column names
}
\description{
The function checks if there are numeric columns which are not
standard integration matrix columns, if there are returns their names.
}
\keyword{internal}
