% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISADataFrame.R
\name{.find_nonNumData}
\alias{.find_nonNumData}
\title{Internal helper function to find the indexes of detected non numeric columns.}
\usage{
.find_nonNumData(x)
}
\arguments{
\item{x}{an ISADataFrame object}
}
\value{
a numeric vector
}
\description{
Internal helper function to find the indexes of detected non numeric columns.
}
