% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.find_unique_max}
\alias{.find_unique_max}
\title{Finds a unique maximum value in a vector of numbers.}
\usage{
.find_unique_max(x)
}
\arguments{
\item{x}{A numeric or integer vector}
}
\value{
A single numeric value or an empty numeric vector
}
\description{
Unlike the standard \code{max} function, this function returns a single maximum
value if and only if a maximum exists, more in details that means:
\itemize{
\item It ignores NA values by default
\item If the identified maximum value is not unique in the vector, it returns
an empty vector instead
}
}
\keyword{internal}
