% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.fix_metadata}
\alias{.fix_metadata}
\title{Internal utility function to use for fixing the metadata attribute after
structural changes to an ISADataFrame object.}
\usage{
.fix_metadata(x)
}
\arguments{
\item{x}{An ISADataFrame object}
}
\value{
The same ISADataFrame as input with fixed metadata
}
\description{
Internal utility function to use for fixing the metadata attribute after
structural changes to an ISADataFrame object.
}
