% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.generate_filename}
\alias{.generate_filename}
\title{Generates a file name for recalibration maps.}
\usage{
.generate_filename()
}
\value{
A string
}
\description{
Unique file names include current date and timestamp.
}
\keyword{internal}
