% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.identify_independent_samples}
\alias{.identify_independent_samples}
\title{Identifies independent samples and separates the joined_df in
collisions and non-collisions}
\usage{
.identify_independent_samples(joined_df)
}
\arguments{
\item{joined_df}{The joined tibble obtained via \code{.join_matrix_af}}
}
\value{
A named list containing the splitted joined_df for collisions and
non-collisions
}
\description{
Identifies independent samples and separates the joined_df in
collisions and non-collisions
}
\keyword{internal}
