% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.import_stats_iss}
\alias{.import_stats_iss}
\title{Imports all found Vispa2 stats files.}
\usage{
.import_stats_iss(association_file)
}
\arguments{
\item{association_file}{The association file}
}
\value{
A list with the imported stats and a report of imported files. If no
files were imported returns NULL instead
}
\description{
Imports all found Vispa2 stats files.
}
\keyword{internal}
