% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.import_type}
\alias{.import_type}
\title{Internal function for parallel import of a single quantification type files.}
\usage{
.import_type(q_type, files, workers)
}
\arguments{
\item{q_type}{The quantification type (single string)}

\item{files}{Files_found table were absolute paths of chosen files are stored}

\item{workers}{Number of parallel workers}
}
\value{
A single tibble with all data from matrices of same quantification
type in tidy format
}
\description{
Internal function for parallel import of a single quantification type files.
}
