% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.interactive_select_projects_import}
\alias{.interactive_select_projects_import}
\title{Allows the user to choose interactively the projects
to consider for import.}
\usage{
.interactive_select_projects_import(association_file)
}
\arguments{
\item{association_file}{The tibble representing the imported association file}
}
\value{
A modified version of the association file where only selected
projects are present
}
\description{
Allows the user to choose interactively the projects
to consider for import.
}
\keyword{internal}
