% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.is_annotated}
\alias{.is_annotated}
\title{Checks if the integration matrix is annotated or not.}
\usage{
.is_annotated(x)
}
\arguments{
\item{x}{A data.frame}
}
\value{
A logical value
}
\description{
Checks if the integration matrix is annotated or not.
}
\keyword{internal}
