% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.iss_import_widget}
\alias{.iss_import_widget}
\title{Builds the html widget for the iss_import.}
\usage{
.iss_import_widget(report)
}
\arguments{
\item{report}{Table obtained via \code{import_stats_iss}}
}
\value{
A widget
}
\description{
Builds the html widget for the iss_import.
}
\keyword{internal}
