% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.join_and_aggregate}
\alias{.join_and_aggregate}
\title{Performs eventual join with stats and aggregation.}
\usage{
.join_and_aggregate(association_file, stats, grouping_keys)
}
\arguments{
\item{association_file}{The imported association file}

\item{stats}{The imported stats df}

\item{grouping_keys}{The names of the columns to group by}
}
\value{
A tibble
}
\description{
Performs eventual join with stats and aggregation.
}
\keyword{internal}
