% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.join_matrix_af}
\alias{.join_matrix_af}
\title{Produces a joined tibble between the sequence count matrix and the
association file}
\usage{
.join_matrix_af(seq_count_df, association_file, date_col)
}
\arguments{
\item{seq_count_df}{The sequence count tibble}

\item{association_file}{The association file tibble}

\item{date_col}{The date column chosen}
}
\value{
A tibble
}
\description{
Produces a joined tibble between the sequence count matrix and the
association file
}
