% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.keep_central}
\alias{.keep_central}
\title{Internal function for keep criteria "keep_central".}
\usage{
.keep_central(x, inverted = TRUE)
}
\arguments{
\item{x}{A tibble or a vector of numbers}

\item{inverted}{Return the values to drop instead of the ones to keep?}
}
\value{
If x is tibble returns a tibble otherwise returns a vector
}
\description{
Returns the central value in a set to keep or returns values to drop if
inverted is TRUE.
}
\keyword{internal}
