% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.keep_max_value}
\alias{.keep_max_value}
\title{Internal function for keep criteria "max_value".}
\usage{
.keep_max_value(x, second_choice, inverted = TRUE)
}
\arguments{
\item{x}{A tibble or a vector of numbers}

\item{second_choice}{Second criteria to use if "max_value" is not applicable}

\item{inverted}{Return the values to drop instead of the ones to keep?}
}
\value{
If x is tibble returns a tibble otherwise returns a vector
}
\description{
Returns the value in a set which has (or is) the max "Value" or returns
values to drop if inverted is TRUE. If values aren't comparable (because
they're equal), a second criteria is used for selection.
}
\keyword{internal}
