% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.locus_distance}
\alias{.locus_distance}
\title{Internal function that computes distance between two loci.}
\usage{
.locus_distance(x, y)
}
\arguments{
\item{x}{First locus}

\item{y}{Second locus}
}
\value{
The distance as the absolute value of x-y
}
\description{
Internal function that computes distance between two loci.
}
\keyword{internal}
