% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.lookup_matrices}
\alias{.lookup_matrices}
\title{Looks up matrices to import given the association file and the
root of the file system.}
\usage{
.lookup_matrices(association_file, quantification_type, matrix_type)
}
\arguments{
\item{association_file}{Tibble representing the association file}

\item{quantification_type}{The type of quantification matrices to look for
(one in \code{quantification_types()})}

\item{matrix_type}{The matrix_type to lookup (one between "annotated" or
"not_annotated")}
}
\value{
A tibble containing all found files, including duplicates and missing
}
\description{
Looks up matrices to import given the association file and the
root of the file system.
}
\keyword{internal}
