% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.lookup_matrices_auto}
\alias{.lookup_matrices_auto}
\title{Looks up matrices to import given the association file and the root of the
file system.}
\usage{
.lookup_matrices_auto(
  association_file,
  quantification_type,
  matrix_type,
  patterns,
  matching_opt
)
}
\arguments{
\item{association_file}{Tibble representing the association file}

\item{quantification_type}{The type of quantification matrices to look for
(one in \code{quantification_types()})}

\item{matrix_type}{The matrix_type to lookup (one between "annotated" or
"not_annotated")}

\item{patterns}{A character vector of patterns to be matched}

\item{matching_opt}{A single character representing the matching option (one
of "ANY", "ALL" or "OPTIONAL")}
}
\value{
A tibble containing all found files, including duplicates and missing
}
\description{
Looks up matrices to import given the association file and the root of the
file system.
}
