% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.manage_anomalies_auto}
\alias{.manage_anomalies_auto}
\title{Manages anomalies for files found.}
\usage{
.manage_anomalies_auto(files_found)
}
\arguments{
\item{files_found}{The tibble obtained via calling \code{.lookup_matrices_auto}}
}
\value{
A tibble containing for each project, pool and quantification type
the files chosen
}
\description{
The function manages anomalies found for files after scanning appropriate
folders by:
\itemize{
\item Removing files not found (files for which Files_count$Found == 0 and Path
is NA) and printing a message to notify user
\item Removing duplicates
}
}
