% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.manage_anomalies_interactive}
\alias{.manage_anomalies_interactive}
\title{Manages anomalies for files found.}
\usage{
.manage_anomalies_interactive(files_found)
}
\arguments{
\item{files_found}{The tibble obtained via calling \code{.lookup_matrices}}
}
\value{
A tibble containing for each project, pool and quantification type
the files chosen (ideally 1 for each quantification type if found, no more
than 1 per type)
}
\description{
The function manages anomalies found for files after scanning appropriate
folders by:
\itemize{
\item Removing files not found (files for which Files_count$Found == 0 and Path
is NA) and printing a message to notify user
\item Removing duplicates by asking the user which files to keep for each
quantification type, pool and project
}
}
