% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.manage_association_file}
\alias{.manage_association_file}
\title{Helper function to be used when importing matrices in parallel.}
\usage{
.manage_association_file(association_file, root, padding, format)
}
\arguments{
\item{association_file}{Either the path to the association file or the tibble
representing the imported association file (done via
\code{import_association_file})}

\item{root}{If \code{association_file} is the path to the file, root is a single
string holding the path to the root folder, otherwise root is \code{NULL}}

\item{padding}{The padding for TimePoint field}

\item{format}{The date format of date columns}
}
\value{
A list of two elements: the first element is an updated version of
the association file with NAs removed, the second element is a widget showing
results of alignment checks.
}
\description{
Helper function to be used when importing matrices in parallel.
}
