% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.min_var_set}
\alias{.min_var_set}
\title{Minimal association_file variable set.}
\usage{
.min_var_set()
}
\value{
A character vector
}
\description{
Contains the names of the columns of the association file that are a minimum
requirement to perform aggregation.
}
\keyword{internal}
