% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.obtain_nested}
\alias{.obtain_nested}
\title{Returns a polished version of collisions table for analysis.}
\usage{
.obtain_nested(collisions)
}
\arguments{
\item{collisions}{The collisions table obtained via
\code{.identify_independent_samples}}
}
\value{
A nested tibble
}
\description{
Polishing consists in nesting all information relative to a single
integration, in other words, obtaining a linked data frame for each
integration.
}
\keyword{internal}
