% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.parallel_import_merge}
\alias{.parallel_import_merge}
\title{Internal function for importing all files for each quantification type.}
\usage{
.parallel_import_merge(files_to_import, workers)
}
\arguments{
\item{files_to_import}{The tibble containing the files to import}

\item{workers}{Number of parallel workers}
}
\value{
A named list of tibbles
}
\description{
Internal function for importing all files for each quantification type.
}
\keyword{internal}
