% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.pattern_matching}
\alias{.pattern_matching}
\title{Internal function to match user defined patterns on a vector
of file names.}
\usage{
.pattern_matching(filenames, patterns)
}
\arguments{
\item{filenames}{A character vector of file names}

\item{patterns}{A character vector of patterns to be matched}
}
\value{
A tibble
}
\description{
For each pattern specified by the user, the function tries to find a match
on all the file names and combines the results as a tibble in which column
names are the patterns and the values are TRUE if the pattern matched on the
element at that index or FALSE otherwise.
}
\keyword{internal}
