% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.pool_choices_IN}
\alias{.pool_choices_IN}
\title{Simple helper interal function to handle user input for actual pool choices.}
\usage{
.pool_choices_IN(indexes)
}
\arguments{
\item{indexes}{A vector of integer indexes available}
}
\value{
The user selection as a numeric vector
}
\description{
Simple helper interal function to handle user input for actual pool choices.
}
