% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.process_collisions}
\alias{.process_collisions}
\title{Internal function to process collisions on multiple integrations,
parallelized.}
\usage{
.process_collisions(collisions, date_col, reads_ratio)
}
\arguments{
\item{collisions}{The collision data frame}

\item{date_col}{The date column to consider}

\item{reads_ratio}{The value of the ratio between sequence count values to
check}
}
\value{
A list containing the updated collisions, a numeric value
representing the number of integrations removed and a numeric value
representing the number of integrations reassigned
}
\description{
Internal function to process collisions on multiple integrations,
parallelized.
}
\keyword{internal}
