% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-widgets.R
\name{.recalibr_map_widget}
\alias{.recalibr_map_widget}
\title{Builds the html widget for the recalibration map.}
\usage{
.recalibr_map_widget(map_rec)
}
\arguments{
\item{map_rec}{Table obtained via \code{.sliding_window}}
}
\value{
A widget
}
\description{
Builds the html widget for the recalibration map.
}
\keyword{internal}
