% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.sliding_window}
\alias{.sliding_window}
\title{Internal function that implements the sliding window algorithm.}
\usage{
.sliding_window(x, threshold, keep_criteria)
}
\arguments{
\item{x}{An integration matrix subset (see description)}

\item{threshold}{The numeric value representing the threshold for selection}

\item{keep_criteria}{The character vector with selecting criterias}
}
\value{
A modified tibble with a number of rows less or equal than x.
}
\description{
\strong{NOTE: this function is meant to be called on a SINGLE GROUP, meaning a
subset of an integration matrix in which all rows share the same chr and same
strand.}
Also note that is better to call this function on a group that has all
distinct integration_locus values (this is ensured by calling function) and
also it's not possible to call this function on groups with a single row.
}
\keyword{internal}
