% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.stats_columns_min}
\alias{.stats_columns_min}
\title{Minimal stats column set.}
\usage{
.stats_columns_min()
}
\value{
A character vector
}
\description{
Contains the name of the columns that are a minimum requirement for
aggregation.
}
\keyword{internal}
