% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.stats_report}
\alias{.stats_report}
\title{Finds automatically the path on disk to each stats file.}
\usage{
.stats_report(association_file)
}
\arguments{
\item{association_file}{The association file}
}
\value{
A tibble with ProjectID and the absolute path on disk to each file
if found
}
\description{
Finds automatically the path on disk to each stats file.
}
\keyword{internal}
