% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.summary_collisions_widget}
\alias{.summary_collisions_widget}
\title{Builds the html widget for the summary table.}
\usage{
.summary_collisions_widget(
  removed,
  reassigned,
  summary,
  tot_rows,
  collision_rows
)
}
\arguments{
\item{removed}{Number of removed collisions}

\item{reassigned}{Number of re-assigned collisions}

\item{summary}{Summary table}

\item{tot_rows}{Total number rows of sequence count matrix before processing}

\item{collision_rows}{Total number of rows of collisions}
}
\value{
A widget
}
\description{
Builds the html widget for the summary table.
}
