% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.summary_table}
\alias{.summary_table}
\title{Internal function to obtain a summary table after collision processing.}
\usage{
.summary_table(before, after, association_file)
}
\arguments{
\item{before}{The joined table before collision removing
(obtained via \code{.join_matrix_af})}

\item{after}{The final matrix obtained after collision processing}

\item{association_file}{The association file}
}
\value{
A tibble with a summary containing for each SubjectID the number of
integrations found before and after, the sum of the value of the sequence
count for each subject before and after and the corresponding deltas.
}
\description{
Internal function to obtain a summary table after collision processing.
}
