% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.trace_anomalies}
\alias{.trace_anomalies}
\title{Updates a files_found tibble with Files_count and Anomalies column.}
\usage{
.trace_anomalies(lups)
}
\arguments{
\item{lups}{A files_found tibble obtained in a lookup function. Must contain
the Files column (nested table quantification type and files)}
}
\value{
Updated files_found with Anomalies and Files_count columns
}
\description{
Updates a files_found tibble with Files_count and Anomalies column.
}
