% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.update_af_after_alignment}
\alias{.update_af_after_alignment}
\title{Updates the association file after the alignment check.}
\usage{
.update_af_after_alignment(as_file, checks, root)
}
\arguments{
\item{as_file}{The tibble representing the read association_file}

\item{checks}{The tibble representing the results
of \code{.check_file_system_alignment}}

\item{root}{The root folder}
}
\value{
An updated association file with absolute paths
}
\description{
The function checks if there are missing folders and updates the association
file by adding a column \code{Path} where the absolute path on disk for the
project and pool is found, if no path is found NA is inserted instead.
}
