% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.update_as_option}
\alias{.update_as_option}
\title{Updates files_found tibble according to pattern and matching options.}
\usage{
.update_as_option(files_nested, p_matches, matching_opt)
}
\arguments{
\item{files_nested}{The tibble containing Quantification_type and Files_found
columns relative to a single project/pool pair}

\item{p_matches}{The tibble representing the pattern matchings resulting from
\code{pattern_matching}}

\item{matching_opt}{The matching option}
}
\value{
An updated files_found tibble according to the matching option
}
\description{
Updates files_found tibble according to pattern and matching options.
}
\keyword{internal}
