% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.window}
\alias{.window}
\title{Internal for the construction of a window of width equal to 3 rows and
computation of result on those values.}
\usage{
.window(indexes, loci, values, criterias, threshold)
}
\arguments{
\item{indexes}{The indexes of the rows in the window
(named vector, see description)}

\item{loci}{The actual value of the \code{integration_locus} variable
for the rows (named vector, see description)}

\item{values}{The actual value of the \code{Value} variable for the rows
(named vector, see description)}

\item{criterias}{The character vector containing the selection criterias}

\item{treshold}{The numeric value representing the threshold for selection}
}
\value{
A named list with the indexes of rows to drop, the index
of the row to collapse on and the value to assign.
If no row needs to be dropped returns NULL instead.
}
\description{
NOTE: to use the function correctly, the parameters
\code{indexes}, \code{loci} and
\code{values} MUST be named vector with these names:
c(first = ..., center = ..., last = ...)
}
\keyword{internal}
