% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.window_slide}
\alias{.window_slide}
\title{Internal that represents the window moving and modifying the
original table as it progresses.}
\usage{
.window_slide(x, start, criterias, threshold)
}
\arguments{
\item{x}{The original tibble (subgroup)}

\item{start}{The starting index from where the window should be built}

\item{criterias}{The character vector with selecting criterias}

\item{threshold}{The numeric value representing the threshold for selection}
}
\value{
A modified tibble with a number of rows less or equal than x.
}
\description{
Internal that represents the window moving and modifying the
original table as it progresses.
}
\keyword{internal}
