% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.wrapper}
\alias{.wrapper}
\title{Wrapper function to call in summarise.}
\usage{
.wrapper(x, lambda, args, namespace, env)
}
\arguments{
\item{x}{The numeric vector to which the lambda is applied}

\item{lambda}{The name of the function to call}

\item{args}{Additional arguments to pass to lambda}

\item{namespace}{The namespace that exports lambda (or NULL)}

\item{env}{The environment where symbols are evaluated}
}
\value{
An atomic type or a list
}
\description{
This function tests for the type of result obtained by calling the lambda on
the grouped \code{Value} column: if its an atomic type (and not a list or a
vector) simply returns the result, otherwise encapsulates the result in a
list to produce a list column that can be unnested with tidyr. Data frames
are automatically converted to tibbles to improve performance.
}
\details{
This function actually improves both performance and flexibility of the
aggregate function since every object of any class can be included in a list
and therefore in the aggregated_value column.
}
\keyword{internal}
