% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.write_recalibr_map}
\alias{.write_recalibr_map}
\title{Tries to write the recalibration map to a tsv file.}
\usage{
.write_recalibr_map(map, file_path)
}
\arguments{
\item{map}{The recalibration map}

\item{file_path}{The file path as a string}
}
\value{
Nothing
}
\description{
Tries to write the recalibration map to a tsv file.
}
