% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{enable_progress_bars}
\alias{enable_progress_bars}
\title{Enable global progress bars for ISAnalytics functions.}
\usage{
enable_progress_bars()
}
\value{
\code{NULL}
}
\description{
This is a simple wrapper around functions from the package
\code{progressr}. To customize the appearance of the progress bar,
please refer to \href{https://progressr.futureverse.org/}{progressr}
documentation.
}
\examples{
enable_progress_bars()
progressr::handlers(global = FALSE) # Deactivate
}
\seealso{
Other Utilities: 
\code{\link{as_sparse_matrix}()},
\code{\link{comparison_matrix}()},
\code{\link{export_ISA_settings}()},
\code{\link{generate_Vispa2_launch_AF}()},
\code{\link{generate_blank_association_file}()},
\code{\link{generate_default_folder_structure}()},
\code{\link{import_ISA_settings}()},
\code{\link{separate_quant_matrices}()},
\code{\link{transform_columns}()}
}
\concept{Utilities}
