% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{generate_blank_association_file}
\alias{generate_blank_association_file}
\title{Create a blank association file.}
\usage{
generate_blank_association_file(path)
}
\arguments{
\item{path}{The path on disk where the file should be written - must be a
file}
}
\value{
\code{NULL}
}
\description{
Produces a blank association file to start using
both VISPA2 and ISAnalytics
}
\examples{
temp <- tempfile()
generate_blank_association_file(temp)
}
\seealso{
Other Utilities: 
\code{\link{as_sparse_matrix}()},
\code{\link{comparison_matrix}()},
\code{\link{generate_Vispa2_launch_AF}()},
\code{\link{generate_default_folder_structure}()},
\code{\link{separate_quant_matrices}()},
\code{\link{transform_columns}()}
}
\concept{Utilities}
