% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{generate_default_folder_structure}
\alias{generate_default_folder_structure}
\title{Title}
\usage{
generate_default_folder_structure(
  type = "correct",
  dir = tempdir(),
  af = "default",
  matrices = "default"
)
}
\arguments{
\item{matrices}{}
}
\value{

}
\description{
Title
}
