% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{HSC_population_plot}
\alias{HSC_population_plot}
\title{Plot of the estimated HSC population size for each patient.}
\usage{
HSC_population_plot(
  estimates,
  project_name,
  timepoints = "Consecutive",
  models = "Mth Chao (LB)"
)
}
\arguments{
\item{estimates}{The estimates data frame, obtained via
\code{\link{HSC_population_size_estimate}}}

\item{project_name}{The project name, will be included in the plot title}

\item{timepoints}{Which time points to plot? One between "All",
"Stable" and "Consecutive"}

\item{models}{Name of the models to plot (as they appear in the column
of the estimates)}
}
\value{
A plot
}
\description{
Plot of the estimated HSC population size for each patient.
}
\examples{
op <- options("ISAnalytics.widgets" = FALSE, "ISAnalytics.verbose" = FALSE)
path_AF <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_correct <- system.file("extdata", "fs.zip", package = "ISAnalytics")
root_correct <- unzip_file_system(root_correct, "fs")
association_file <- import_association_file(path_AF, root_correct,
    dates_format = "dmy"
)
aggregated_meta <- aggregate_metadata(association_file)
matrices <- import_parallel_Vispa2Matrices_auto(
    association_file = association_file, root = NULL,
    quantification_type = c("fragmentEstimate", "seqCount"),
    matrix_type = "annotated", workers = 2, matching_opt = "ANY"
)
agg <- aggregate_values_by_key(
    x = matrices,
    association_file = association_file,
    value_cols = "seqCount"
)
estimate <- HSC_population_size_estimate(
    x = agg,
    metadata = aggregated_meta,
    stable_timepoints = NULL
)
p <- HSC_population_plot(estimate, "PROJECT1")
options(op)
}
\seealso{
Other Plotting functions: 
\code{\link{CIS_volcano_plot}()},
\code{\link{integration_alluvial_plot}()}
}
\concept{Plotting functions}
