% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\name{import_parallel_Vispa2Matrices_auto}
\alias{import_parallel_Vispa2Matrices_auto}
\title{Import integration matrices based on the association file.}
\usage{
import_parallel_Vispa2Matrices_auto(
  association_file,
  root,
  quantification_type,
  matrix_type = "annotated",
  workers = 2,
  patterns = NULL,
  matching_opt = matching_options(),
  tp_padding = 4,
  dates_format = "ymd",
  multi_quant_matrix = FALSE,
  export_widget_path = NULL,
  ...
)
}
\arguments{
\item{association_file}{A single string containing the path to the
association file on disk, or a tibble resulting from the previous call of
\code{import_association_file}}

\item{root}{A single string containing the path to the root folder containing
Vispa2 output. Can be NULL if association_file parameter is a tibble}

\item{quantification_type}{A vector of requested quantification_types. Must
be one in \code{quantification_types()}}

\item{matrix_type}{A single string representing the type of matrices to
be imported. Can only be one in \code{"annotated"} or \code{"not_annotated"}}

\item{workers}{A single number representing the number of parallel workers to
use for the import}

\item{patterns}{A character vector of additional patterns to match on file
names. Please note that patterns must be regular expressions. Can be NULL if
no patterns needs to be matched.}

\item{matching_opt}{A single value between \code{matching_options}}

\item{tp_padding}{Timepoint padding, indicates the number of digits of the
"Timepoint" column once imported. Fills the content with 0s up to the length
specified (ex: 1 becomes 0001 with a tp_padding of 4)}

\item{dates_format}{A single string indicating how dates should be parsed.
Must be a value in: \code{date_formats()}}

\item{multi_quant_matrix}{If set to TRUE will produce a
multi-quantification matrix (data frame) through \code{comparison_matrix}
instead of a list.}

\item{export_widget_path}{A path on disk to save produced widgets or NULL
if the user doesn't wish to save the html file}

\item{...}{Additional arguments to pass to \code{comparison_matrix}}
}
\value{
A named list of tibbles containing data from all imported integration
matrices, divided by quantification type
}
\description{
\lifecycle{maturing}
These functions are designed to import the appropriate
integration matrix files given the association file and the root folder of
the file system where Vispa2 matrices are generated.
}
\details{
Import family functions are designed to work in combination with
Vispa2, for more details on this take a look here:
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5702242/}{VISPA2:
A Scalable Pipeline for High-Throughput Identification
and Annotation of Vector Integration Sites}.\cr
For more details on how to properly use these functions, refer to the
vignette - vignette("how_to_import_functions")
}
\section{Automatic version}{

The automatic version of import_parallel_Vispa2Matrices doesn't interact with
the user directly, for this reason options in this modality are more limited
compared to the interactive version. In automatic version you can't:
\itemize{
\item Choose single projects or pools: to have a selection import
the association file first and filter it according to your needs
before calling the function
(more details on this in the vignette)
\item Choose duplicates: if, after filtering by the specified patterns,
duplicates are found they are automatically ignored
}
}

\section{Interactive version}{

The interactive version of import_parallel_Vispa2Matrices asks user for input
and allows a more detailed choice of projects to import, pools to import and,
if necessary, duplicate files. During the execution, a series of reports is
shown in html format.
}

\examples{
op <- options("ISAnalytics.widgets" = FALSE)
path <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_pth <- system.file("extdata", "fs.zip", package = "ISAnalytics")
root <- unzip_file_system(root_pth, "fs")
matrices <- import_parallel_Vispa2Matrices_auto(
    path, root,
    c("fragmentEstimate", "seqCount"), "annotated", 2, NULL, "ANY",
    dates_format = "dmy"
)
options(op)
}
\seealso{
\code{\link{matching_options}},
\url{https://stringr.tidyverse.org/articles/regular-expressions.html}

Other Import functions: 
\code{\link{import_association_file}()},
\code{\link{import_parallel_Vispa2Matrices_interactive}()},
\code{\link{import_single_Vispa2Matrix}()}
}
\concept{Import functions}
