% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\name{import_parallel_Vispa2Matrices_interactive}
\alias{import_parallel_Vispa2Matrices_interactive}
\title{Import integration matrices based on the association file.}
\usage{
import_parallel_Vispa2Matrices_interactive(
  association_file,
  root,
  quantification_type,
  matrix_type = "annotated",
  workers = 2,
  tp_padding = 4,
  dates_format = "dmy",
  multi_quant_matrix = FALSE,
  ...
)
}
\arguments{
\item{association_file}{A single string containing the path to the
association file on disk, or a tibble resulting from the previous call of
\code{import_association_file}}

\item{root}{A single string containing the path to the root folder containing
Vispa2 output. Can be NULL if association_file parameter is a tibble}

\item{quantification_type}{A vector of requested quantification_types. Must
be one in \code{quantification_types()}}

\item{matrix_type}{A single string representing the type of matrices to
be imported. Can only be one in \code{"annotated"} or \code{"not_annotated"}}

\item{workers}{A single number representing the number of parallel workers to
use for the import}

\item{tp_padding}{Timepoint padding, indicates the number of digits of the
"Timepoint" column once imported. Fills the content with 0s up to the length
specified (ex: 1 becomes 0001 with a tp_padding of 4)}

\item{dates_format}{A single string indicating how dates should be parsed.
Must be a value in: \code{date_formats()}}

\item{multi_quant_matrix}{If set to TRUE will produce a
multi-quantification matrix (data frame) through \code{comparison_matrix}
instead of a list.}

\item{...}{Additional arguments to pass to \code{comparison_matrix}}
}
\value{
A named list of tibbles containing data from all imported integration
matrices, divided by quantification type
}
\description{
\lifecycle{maturing}
These functions are designed to import the appropriate
integration matrix files given the association file and the root folder of
the file system where Vispa2 matrices are generated.
}
\details{
Import family functions are designed to work in combination with
Vispa2, for more details on this take a look here:
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5702242/}{VISPA2:
A Scalable Pipeline for High-Throughput Identification
and Annotation of Vector Integration Sites}.\cr
For more details on how to properly use these functions, refer to the
vignette - vignette("how_to_import_functions")
}
\section{Interactive version}{

The interactive version of import_parallel_Vispa2Matrices asks user for input
and allows a more detailed choice of projects to import, pools to import and,
if necessary, duplicate files. During the execution, a series of reports is
shown in html format.
}

\examples{
\dontrun{
# Can't run because it's interactive and requires user input
matrices <- import_parallel_Vispa2Matrices_interactive(
    association_file,
    root, quantification_type, matrix_type, workers
)
}
}
\seealso{
\code{\link{comparison_matrix}}

Other Import functions: 
\code{\link{import_association_file}()},
\code{\link{import_parallel_Vispa2Matrices_auto}()},
\code{\link{import_single_Vispa2Matrix}()}
}
\concept{Import functions}
