% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{integration_matrices}
\alias{integration_matrices}
\title{Example of imported multi-quantification integration matrices.}
\format{
Data frame with 1547 rows and 8 columns
\describe{
\item{chr}{The chromosome number (as character)}
\item{integration_locus}{Number of the base at
which the viral insertion occurred}
\item{strand}{Strand of the integration}
\item{GeneName}{Symbol of the closest gene}
\item{GeneStrand}{Strand of the closest gene}
\item{CompleteAmplificationID}{Unique sample identifier}
\item{seqCount}{Value of the sequence count quantification}
\item{fragmentEstimate}{Value of the fragment estimate quantification}
}
}
\usage{
data("integration_matrices")
}
\description{
The data was obtain by random sampling and anonimization of real research
data.
}
\keyword{datasets}
