% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{is.ISADataFrame}
\alias{is.ISADataFrame}
\title{Is the object an ISADataFrame?}
\usage{
is.ISADataFrame(x)
}
\arguments{
\item{x}{an object}
}
\value{
TRUE or FALSE
}
\description{
Is the object an ISADataFrame?
}
\examples{
is.ISADataFrame(1:10)
}
