% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISADataFrame.R
\name{ISADataFrame}
\alias{ISADataFrame}
\title{Helper function to obtain ISADataFrame object.}
\usage{
ISADataFrame(x, metadata = character(), try.correct = TRUE)
}
\arguments{
\item{x}{a named list, a tibble or a data.frame}

\item{metadata}{the metadata fields that are present in the table (should be
all variables that are not mandatory and are not experimental data)}

\item{try.correct}{if set to TRUE is able to fix minor issues}
}
\value{
a properly built ISADataFrame
}
\description{
This function is intended to be used interactively and should be
used to build correct ISADataFrames.
If called with parameter \code{try.correct = TRUE} the function is able to catch
and correct minor issues such as:
\itemize{
\item When provided with a named list as a parameter, if the elements do not have
the same length the shortest are filled with NAs to match the longest element
\item When there are metadata attributes declared that are not present in the
data frame they're removed
\item When non-numeric columns are detected but are not declared as metadata
they're added to the metadata attribute.
}

Errors will be thrown in at least 2 cases:
\itemize{
\item The mandatoryVars are not included in the data frame
\item There are no experimental data columns. Note that experimental data columns
must be numeric.
}
}
\examples{
aListWithSomeIssues <- list(chr = c(as.character(1:10)),
integration_locus = runif(10, min = 100, max = 10000),
strand = sample(c("+", "-"), 10, replace = TRUE),
meta1 = rep_len("m1", 10),
nonNumericdata = rep_len("random", 10),
exp_1 = runif(5, min = 0, max = 10000),
exp_2 = runif(10, min = 0, max = 10000),
exp_3 = runif(8, min = 0, max = 10000))

isadf <- ISADataFrame(aListWithSomeIssues, metadata = c("meta1"),
try.correct = TRUE)
head(isadf)
}
