% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISAnalytics-deprecated.R
\name{ISAnalytics-deprecated}
\alias{ISAnalytics-deprecated}
\title{Deprecated functions in package \pkg{ISAnalytics}.}
\description{
These functions are provided for compatibility
with older versions of \sQuote{ISAnalytics} only,
and will be defunct at the next release.
}
\details{
The following functions are deprecated and will be made defunct;
use the replacement indicated below:
\itemize{
\item import_parallel_Vispa2Matrices_auto:
\code{\link{import_parallel_Vispa2Matrices}}
\item import_parallel_Vispa2Matrices_interactive:
\code{\link{import_parallel_Vispa2Matrices}}
\item unzip_file_system:
\code{\link{generate_default_folder_structure}}
\item cumulative_count_union:
\code{\link{cumulative_is}}
}
}
\keyword{internal}
