% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISAnalytics-package.R
\docType{package}
\name{ISAnalytics-package}
\alias{ISAnalytics}
\alias{ISAnalytics-package}
\title{ISAnalytics: Analyze gene therapy vector insertion sites data identified from genomics next generation sequencing reads for clonal tracking studies}
\description{
In gene therapy, stem cells are modified using viral vectors to deliver the therapeutic transgene and replace functional properties since the genetic modification is stable and inherited in all cell progeny. The retrieval and mapping of the sequences flanking the virus-host DNA junctions allows the identification of insertion sites (IS), essential for monitoring the evolution of genetically modified cells in vivo. A comprehensive toolkit for the analysis of IS is required to foster clonal trackign studies and supporting the assessment of safety and long term efficacy in vivo. This package is aimed at (1) supporting automation of IS workflow, (2) performing base and advance analysis for IS tracking (clonal abundance, clonal expansions and statistics for insertional mutagenesis, etc.), (3) providing basic biology insights of transduced stem cells in vivo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com//calabrialab/isanalytics}
  \item Report bugs at \url{https://github.com//calabrialab/isanalytics}
}

}
\author{
\strong{Maintainer}: Andrea Calabria \email{calabria.andrea@hsr.it}

Authors:
\itemize{
  \item Giulio Spinozzi \email{spinozzi.giulio@hsr.it}
  \item Giulia Pais \email{giuliapais1@gmail.com}
}

}
\keyword{internal}
