% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{known_clinical_oncogenes}
\alias{known_clinical_oncogenes}
\title{Known clinical oncogenes (for mouse and human).}
\usage{
known_clinical_oncogenes()
}
\value{
A data frame
}
\description{
Known clinical oncogenes (for mouse and human).
}
\examples{
known_clinical_oncogenes()
}
\seealso{
Other Plotting function helpers: 
\code{\link{clinical_relevant_suspicious_genes}()}
}
\concept{Plotting function helpers}
