% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{mandatory_IS_vars}
\alias{mandatory_IS_vars}
\title{Names of mandatory variables for an integration matrix.}
\usage{
mandatory_IS_vars(include_types = FALSE)
}
\arguments{
\item{include_types}{If set to \code{TRUE} returns both the names and the types
associated, otherwise returns only a character vector of names}
}
\value{
A character vector or a data frame
}
\description{
Contains the names of the columns that need to be present in order for a
tibble to be considered an integration matrix. To set
custom mandatory variables see the function
\code{\link{set_mandatory_IS_vars}}.
}
\examples{
# Names only
mandatory_IS_vars()

# Names and types
mandatory_IS_vars(TRUE)
}
