% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISADataFrame.R
\name{mandatoryVars}
\alias{mandatoryVars}
\title{Gets the value of the attribute mandatoryVars.}
\usage{
mandatoryVars(x)
}
\arguments{
\item{x}{an ISADataFrame object}
}
\value{
a character vector
}
\description{
Gets the value of the attribute mandatoryVars.
}
\examples{
\dontrun{
mandatory <- mandatoryVars(isadf)
}
}
