% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\name{matching_options}
\alias{matching_options}
\title{Possible choices for the \code{matching_opt} parameter.}
\usage{
matching_options()
}
\value{
A vector of characters for matching_opt
}
\description{
These are all the possible values for the \code{matching_opt} parameter in
\code{import_parallel_vispa2Matrices_auto}.
}
\details{
The values "ANY", "ALL" and "OPTIONAL", represent how the patterns
should be matched, more specifically
\itemize{
\item ANY = look only for files that match AT LEAST one of the
patterns specified
\item ALL = look only for files that match ALL of the patterns specified
\item OPTIONAL = look preferentially for files that match, in order, all
patterns or any pattern and if no match is found return what is found (keep
in mind that duplicates are discarded in automatic mode)
}
}
\examples{
opts <- matching_options()
}
\seealso{
\code{\link{import_parallel_Vispa2Matrices_auto}}

Other Import functions helpers: 
\code{\link{annotation_issues}()},
\code{\link{quantification_types}()}
}
\concept{Import functions helpers}
