% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISADataFrame.R
\name{metadata}
\alias{metadata}
\title{Gets the value of the attribute metadata.}
\usage{
metadata(x)
}
\arguments{
\item{x}{an ISADataFrame object}
}
\value{
a character vector
}
\description{
Gets the value of the attribute metadata.
}
\examples{
 \dontrun{
meta <- metadata(isadf)
}
}
