% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{NGSdataExplorer}
\alias{NGSdataExplorer}
\title{Launch the shiny application NGSdataExplorer.}
\usage{
NGSdataExplorer()
}
\value{
Nothing
}
\description{
Launch the shiny application NGSdataExplorer.
}
\examples{
\dontrun{NGSdataExplorer()}
}
