% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISADataFrame.R
\name{print.ISADataFrame}
\alias{print.ISADataFrame}
\title{Printing ISADataFrames}
\usage{
\method{print}{ISADataFrame}(x, ...)
}
\arguments{
\item{x}{an ISADataFrame object}

\item{...}{optional arguments to print}
}
\description{
Printing ISADataFrames
}
\examples{
\dontrun{
print(isadf)
}
}
