% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proto_oncogenes}
\alias{proto_oncogenes}
\alias{tumor_suppressors}
\title{Data frames for proto-oncogenes (human and mouse)
and tumor-suppressor genes from UniProt.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 569 rows and 13 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 523 rows and 13 columns.
}
\usage{
data("proto_oncogenes")

data("tumor_suppressors")
}
\description{
The file is simply a result of a research with the keywords
"proto-oncogenes" and "tumor suppressor" for the target genomes
on UniProt database.
}
\section{Functions}{
\itemize{
\item \code{tumor_suppressors}: Data frame for tumor suppressor genes

}}
\keyword{datasets}
