% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\name{quantification_types}
\alias{quantification_types}
\title{Possible choices for the \code{quantification_type} parameter.}
\usage{
quantification_types()
}
\value{
A vector of characters for quantification types
}
\description{
These are all the possible values for the \code{quantification_type} parameter in
\code{import_parallel_vispa2Matrices_interactive} and
\code{import_parallel_vispa2Matrices_auto}.
}
\details{
The possible values are:
\itemize{
\item fragmentEstimate
\item seqCount
\item barcodeCount
\item cellCount
\item ShsCount
}
}
\examples{
quant_types <- quantification_types()
}
\seealso{
\code{
\link[import_parallel_vispa2Matrices]{
import_parallel_vispa2Matrices_interactive}},
\code{
\link[import_parallel_vispa2Matrices]{
import_parallel_vispa2Matrices_auto}}

Other Import functions helpers: 
\code{\link{matching_options}()}
}
\concept{Import functions helpers}
