% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported-vars.R
\name{reduced_AF_columns}
\alias{reduced_AF_columns}
\title{Names of the columns of the association file to consider for
Vispa2 launch.}
\usage{
reduced_AF_columns()
}
\value{
A character vector
}
\description{
Selection of column names from the association file to be considered for
Vispa2 launch.
NOTE: the \code{TagID} column appears only once but needs to be
repeated twice for generating the launch file. Use the appropriate
function to generate the file automatically.
}
\examples{
reduced_AF_columns()
}
