% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported-vars.R
\name{refGene_table_cols}
\alias{refGene_table_cols}
\title{Required columns for refGene file.}
\usage{
refGene_table_cols()
}
\value{
Character vector of column names
}
\description{
Required columns for refGene file.
}
\examples{
refGene_table_cols()
}
