% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{reset_af_columns_def}
\alias{reset_af_columns_def}
\title{Resets the association file columns definitions to the default value.}
\usage{
reset_af_columns_def()
}
\value{
Nothing
}
\description{
Resets the association file columns definitions to the default value.
}
\examples{
temp_af_cols <- tibble::tribble(
  ~ names, ~ types, ~ transform,
  "ProjectID", "char", NULL,
  "SubjectID", "char", NULL,
  "TimePoint", "int", NULL
)
set_af_columns_def(temp_af_cols)
print(association_file_columns(TRUE))
reset_af_columns_def()
}
\seealso{
\code{\link{association_file_columns}},
\code{\link{set_af_columns_def}}
}
