% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{reset_annotation_IS_vars}
\alias{reset_annotation_IS_vars}
\title{Resets annotation IS vars to the default value.}
\usage{
reset_annotation_IS_vars()
}
\value{
Nothing
}
\description{
Resets annotation IS vars to the default value.
}
\examples{
set_annotation_IS_vars(c(GeneName = "char"))
print(annotation_IS_vars(TRUE))
reset_annotation_IS_vars()
print(annotation_IS_vars())
}
\seealso{
\code{\link{annotation_IS_vars}},
\code{\link{set_annotation_IS_vars}}
}
