% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{reset_mandatory_IS_vars}
\alias{reset_mandatory_IS_vars}
\title{Resets mandatory IS vars to the default value.}
\usage{
reset_mandatory_IS_vars()
}
\value{
Nothing
}
\description{
Resets mandatory IS vars to the default value.
}
\examples{
set_mandatory_IS_vars(c(chr = "char"))
print(mandatory_IS_vars())
reset_mandatory_IS_vars()
print(mandatory_IS_vars())
}
\seealso{
\code{\link{mandatory_IS_vars}},
\code{\link{set_mandatory_IS_vars}}
}
