% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{select.ISADataFrame}
\alias{select.ISADataFrame}
\title{select implementation for ISADataFrame.}
\usage{
\method{select}{ISADataFrame}(.data, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables.}
}
\value{
An ISADataFrame
}
\description{
select implementation for ISADataFrame.
}
\examples{
path <- system.file("extdata", "ex_annotated_ISMatrix.tsv.xz",
    package = "ISAnalytics"
)
isadf1 <- import_single_Vispa2Matrix(path)
isadf1 \%>\% dplyr::select(chr)
}
