% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-functions.R
\name{separate_quant_matrices}
\alias{separate_quant_matrices}
\title{Separate a multiple-quantification matrix into single quantification
matrices.}
\usage{
separate_quant_matrices(
  x,
  fragmentEstimate = "fragmentEstimate",
  seqCount = "seqCount",
  barcodeCount = "barcodeCount",
  cellCount = "cellCount",
  ShsCount = "ShsCount",
  key = c(mandatory_IS_vars(), annotation_IS_vars(), "CompleteAmplificationID")
)
}
\arguments{
\item{x}{Single integration matrix with multiple quantification
value columns, likely obtained via \link{comparison_matrix}.}

\item{fragmentEstimate}{Name of the fragment estimate values column
in input}

\item{seqCount}{Name of the sequence count values column
in input}

\item{barcodeCount}{Name of the barcode count values column
in input}

\item{cellCount}{Name of the cell count values column
in input}

\item{ShsCount}{Name of the shs count values column
in input}

\item{key}{Key columns to perform the joining operation}
}
\value{
A named list of tibbles, where names are quantification types
}
\description{
\lifecycle{stable}
The function separates a single multi-quantification integration
matrix, obtained via \link{comparison_matrix}, into single
quantification matrices as a named list of tibbles.
}
\examples{
data("integration_matrices", package = "ISAnalytics")
separated <- separate_quant_matrices(
    integration_matrices
)
separated
}
\seealso{
\link{quantification_types}

Other Analysis functions: 
\code{\link{CIS_grubbs}()},
\code{\link{comparison_matrix}()},
\code{\link{compute_abundance}()},
\code{\link{cumulative_count_union}()},
\code{\link{is_sharing}()},
\code{\link{sample_statistics}()},
\code{\link{threshold_filter}()},
\code{\link{top_integrations}()}
}
\concept{Analysis functions}
