% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{set_matrix_file_suffixes}
\alias{set_matrix_file_suffixes}
\title{Title}
\usage{
set_matrix_file_suffixes(
  quantification_suffix = list(seqCount = "seqCount", fragmentEstimate =
    "fragmentEstimate", barcodeCount = "barcodeCount", cellCount = "cellCount", ShsCount
    = "ShsCount"),
  annotation_suffix = list(annotated = ".no0.annotated", not_annotated = ""),
  file_ext = "tsv.gz",
  glue_file_spec = "{quantification_suffix}_matrix{annotation_suffix}.{file_ext}"
)
}
\arguments{
\item{glue_file_spec}{}
}
\value{

}
\description{
Title
}
