% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISADataFrame.R
\name{validate_ISADataFrame}
\alias{validate_ISADataFrame}
\title{Validator for ISADataFrame objects.}
\usage{
validate_ISADataFrame(x)
}
\arguments{
\item{x}{the ISADataFrame object to validate}
}
\value{
TRUE if all checks pass, error otherwise
}
\description{
This is a devel function and should \strong{not} be used interactively.
The validator takes an ISADataFrame as input to check if the object was built correctly. More specifically:\cr
\itemize{
\item Checks if the data frame contains the mandatory variables specified by the \code{mandatoryVars} attribute
\item Checks if the data frame contains the metadata variables specified by the \code{metadata} attribute (if not, generates only a warning)
\item Checks if there is at least one experimental data column: a column is considered experimental data if it's name is not contained
both in \code{mandatoryVars} and \code{metadata} and the column contains numeric values.
}

If all checks pass the function returns TRUE, otherwise some kind of error is shown.
}
