% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate-functions.R
\name{aggregate_metadata}
\alias{aggregate_metadata}
\title{Performs aggregation on metadata contained in the association file.}
\usage{
aggregate_metadata(
  association_file,
  grouping_keys = c("SubjectID", "CellMarker", "Tissue", "TimePoint"),
  import_stats = TRUE
)
}
\arguments{
\item{association_file}{The imported association file
(via \code{import_association_file})}

\item{grouping_keys}{A character vector of column names to form a group}

\item{import_stats}{Should Vispa2 stats files be imported and included?}
}
\value{
A tibble
}
\description{
\lifecycle{experimental}
Groups metadata by grouping_keys and returns a summary of info for each
group. For more details on how to use this function:
\code{vignette("Working with aggregate functions", package = "ISAnalytics")}
}
\examples{
op <- options("ISAnalytics.widgets" = FALSE, "ISAnalytics.verbose" = FALSE)
path_AF <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_correct <- system.file("extdata", "fs.zip", package = "ISAnalytics")
root_correct <- unzip_file_system(root_correct, "fs")
association_file <- import_association_file(path_AF, root_correct,
    dates_format = "dmy"
)
aggregated_meta <- aggregate_metadata(association_file, import_stats = FALSE)
options(op)
}
\seealso{
Other Aggregate functions: 
\code{\link{aggregate_values_by_key}()}
}
\concept{Aggregate functions}
