% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{annotation_IS_vars}
\alias{annotation_IS_vars}
\title{Names of the genomic annotation variables for an integration matrix.}
\usage{
annotation_IS_vars(include_types = FALSE)
}
\arguments{
\item{include_types}{If set to \code{TRUE} returns both the names and the types
associated, otherwise returns only a character vector of names}
}
\value{
A character vector or a data frame
}
\description{
Default values are \verb{GeneName (char), GeneStrand (char)}. To set
custom genomic annotation variables see the function
\code{\link{set_annotation_IS_vars}}.
}
\examples{
# Names only
annotation_IS_vars()

# Names and types
annotation_IS_vars(TRUE)
}
