% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported-vars.R
\name{association_file_columns}
\alias{association_file_columns}
\title{Names of the columns in the association file.}
\usage{
association_file_columns()
}
\value{
A character vector
}
\description{
All the names of the columns present in the association file.
}
\examples{
association_file_columns()
}
