% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-functions.R
\name{CIS_grubbs}
\alias{CIS_grubbs}
\title{Grubbs test for Common Insertion Sites (CIS).}
\usage{
CIS_grubbs(
  x,
  genomic_annotation_file = system.file("extdata", "hg19.refGene.oracle.tsv.xz",
    package = "ISAnalytics"),
  grubbs_flanking_gene_bp = 1e+05,
  threshold_alpha = 0.05,
  add_standard_padjust = TRUE
)
}
\arguments{
\item{x}{An integration matrix, must include the \code{mandatory_IS_vars()}
columns and the \code{annotation_IS_vars()} columns}

\item{genomic_annotation_file}{Database file for gene annotation,
see details}

\item{grubbs_flanking_gene_bp}{Number of base pairs flanking a gene}

\item{threshold_alpha}{Significance threshold}

\item{add_standard_padjust}{Compute the standard padjust?}
}
\value{
A data frame
}
\description{
\lifecycle{experimental}
Statistical approach for the validation of common insertion sites
significance based on the comparison of the integration frequency
at the CIS gene with respect to other genes contained in the
surrounding genomic regions. For more details please refer to
this paper:
\url{https://ashpublications.org/blood/article/117/20/5332/21206/Lentiviral-vector-common-integration-sites-in}
}
\details{
\subsection{Genomic annotation file}{

This file is a data base, or more simply a .tsv file to import, with
genes annotation for the specific genome. The annotations for the
human genome (hg19) is already included in this package.
If for any reason the user is performing an analysis on another genome,
this file needs to be changed respecting the USCS Genome Browser
format, meaning the input file headers should be:\preformatted{## name2, chrom, strand 
##  min_txStart, max_txEnd, minmax_TxLen 
##  average_TxLen, name, min_cdsStart 
##  max_cdsEnd, minmax_CdsLen, average_CdsLen
}
}
}
\examples{
op <- options(ISAnalytics.widgets = FALSE)

path_AF <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_correct <- system.file("extdata", "fs.zip",
    package = "ISAnalytics"
)
root_correct <- unzip_file_system(root_correct, "fs")

matrices <- import_parallel_Vispa2Matrices_auto(
    association_file = path_AF, root = root_correct,
    quantification_type = c("seqCount", "fragmentEstimate"),
    matrix_type = "annotated", workers = 2, patterns = NULL,
    matching_opt = "ANY",
    dates_format = "dmy"
)

cis <- CIS_grubbs(matrices$seqCount)

options(op)
}
\seealso{
Other Analysis functions: 
\code{\link{comparison_matrix}()},
\code{\link{compute_abundance}()},
\code{\link{cumulative_count_union}()},
\code{\link{sample_statistics}()},
\code{\link{separate_quant_matrices}()},
\code{\link{threshold_filter}()},
\code{\link{top_integrations}()}
}
\concept{Analysis functions}
