% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{CIS_volcano_plot}
\alias{CIS_volcano_plot}
\title{Trace volcano plot for computed CIS data.}
\usage{
CIS_volcano_plot(
  x,
  onco_db_file = system.file("extdata", "201806_uniprot-Proto-oncogene.tsv.xz", package
    = "ISAnalytics"),
  tumor_suppressors_db_file = system.file("extdata",
    "201806_uniprot-Tumor-suppressor.tsv.xz", package = "ISAnalytics"),
  species = "human",
  known_onco = known_clinical_oncogenes(),
  suspicious_genes = clinical_relevant_suspicious_genes(),
  significance_threshold = 0.05,
  annotation_threshold_ontots = 0.1,
  facet_rows = NULL,
  facet_cols = NULL
)
}
\arguments{
\item{x}{Either a simple integration matrix or a data frame resulting
from the call to \link{CIS_grubbs} with \code{add_standard_padjust = TRUE}}

\item{onco_db_file}{Uniprot file for proto-oncogenes (see details)}

\item{tumor_suppressors_db_file}{Uniprot file for tumor-suppressor genes}

\item{species}{One between "human", "mouse" and "all"}

\item{known_onco}{Data frame with known oncogenes. See details.}

\item{suspicious_genes}{Data frame with clinical relevant suspicious
genes. See details.}

\item{significance_threshold}{The significance threshold}

\item{annotation_threshold_ontots}{Value above which genes are annotated}

\item{facet_rows}{Column names to pass to the faceting function as rows
(NULL for no faceting)}

\item{facet_cols}{Column names to pass to the faceting function as columns
(NULL for no faceting)}
}
\value{
A plot
}
\description{
\lifecycle{experimental}
Traces a volcano plot for IS frequency and CIS results.
}
\details{
\subsection{Input data frame}{

Users can supply as \code{x} either a simple integration matrix or a
data frame resulting from the call to \link{CIS_grubbs}
with \code{add_standard_padjust = TRUE}. In the first case an internal call to
the function \code{CIS_grubbs} is performed.
}

\subsection{Oncogene and tumor suppressor genes files}{

These files are included in the package for user convenience and are
simply UniProt files with gene annotations for human and mouse.
For more details on how this files were generated use the help \code{?filname}
function.
}

\subsection{Known oncogenes}{

The default values are contained in a data frame exported by this package,
it can be accessed by doing:\if{html}{\out{<div class="r">}}\preformatted{head(known_clinical_oncogenes())
}\if{html}{\out{</div>}}\preformatted{## # A tibble: 5 x 2
##   GeneName KnownClonalExpansion
##   <chr>    <lgl>               
## 1 MECOM    TRUE                
## 2 CCND2    TRUE                
## 3 TAL1     TRUE                
## 4 LMO2     TRUE                
## 5 HMGA2    TRUE
}

If the user wants to change this parameter the input data frame must
preserve the column structure. The same goes for the \code{suspicious_genes}
parameter (DOIReference column is optional):\if{html}{\out{<div class="r">}}\preformatted{head(clinical_relevant_suspicious_genes())
}\if{html}{\out{</div>}}\preformatted{## # A tibble: 6 x 3
##   GeneName ClinicalRelevance DOIReference                                
##   <chr>    <lgl>             <chr>                                       
## 1 DNMT3A   TRUE              https://doi.org/10.1182/blood-2018-01-829937
## 2 TET2     TRUE              https://doi.org/10.1182/blood-2018-01-829937
## 3 ASXL1    TRUE              https://doi.org/10.1182/blood-2018-01-829937
## 4 JAK2     TRUE              https://doi.org/10.1182/blood-2018-01-829937
## 5 CBL      TRUE              https://doi.org/10.1182/blood-2018-01-829937
## 6 TP53     TRUE              https://doi.org/10.1182/blood-2018-01-829937
}
}
}
\examples{
op <- options(ISAnalytics.widgets = FALSE)

path_AF <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_correct <- system.file("extdata", "fs.zip",
    package = "ISAnalytics"
)
root_correct <- unzip_file_system(root_correct, "fs")

matrices <- import_parallel_Vispa2Matrices_auto(
    association_file = path_AF, root = root_correct,
    quantification_type = c("seqCount", "fragmentEstimate"),
    matrix_type = "annotated", workers = 2, patterns = NULL,
    matching_opt = "ANY"
)

cis <- CIS_grubbs(matrices$seqCount)
plot <- CIS_volcano_plot(cis)
options(op)
}
\concept{Plotting functions}
