% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-functions.R
\name{compute_abundance}
\alias{compute_abundance}
\title{Computes the abundance of every integration in the sample.}
\usage{
compute_abundance(x, percentage = TRUE)
}
\arguments{
\item{x}{An integration matrix}

\item{percentage}{Add abundance as percentage?}
}
\value{
An integration matrix
}
\description{
Abundance is obtained for every row by calculating the ratio
between the single value and the total value for the sample.
}
\examples{
path <- system.file("extdata", "ex_annotated_ISMatrix.tsv.xz",
package = "ISAnalytics")
matrix <- import_single_Vispa2Matrix(path)
abundance <- compute_abundance(matrix)
}
\concept{Analysis functions}
