% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalibration-functions.R
\name{compute_near_integrations}
\alias{compute_near_integrations}
\title{Scans input matrix to find and merge near integration sites.}
\usage{
compute_near_integrations(
  x,
  threshold = 4,
  keep_criteria = c("max_value", "keep_central", "keep_first"),
  strand_specific = TRUE
)
}
\arguments{
\item{x}{A single integration matrix}

\item{threshold}{A single numeric value, represents the number of bases
under which two integrations are deemed "near"}

\item{keep_criteria}{While scanning, which integration should be kept?
The 3 possible choices for this parameter are:
\itemize{
\item "max_value": keep the integration site which has the highest value
(and collapse other values on that integration). It is highly reccomended,
if you choose this criteria, to express a second preference in case
values can't be compared, for example c("max_value", "keep_first")
\item "keep_first": keeps the first integration in the considered
window of 3
\item "keep_central": keeps the central integration in the considered
window of 3
}}

\item{strand_specific}{Should strand be considered? If yes,
for example these two integration sites
c(chr = "1", strand = "+", integration_locus = 14568) and
c(chr = "1", strand = "-", integration_locus = 14568) are considered
different and not grouped together.}
}
\value{
An integration matrix with same or less number of rows
}
\description{
This function scans the input integration matrix to detect eventual
integration sites that are too "near" to each other and merges them
into single integration sites that have a value equal to the sum
of values.
}
\details{
The whole matrix is scanned with a sliding window mechanism,
a window being a group of 3 integration sites (or rows which share
the same chromosome and optionally strand). The distance between
two integration sites that share the same chr (and strand optionally),
is calculated in the window by comparing the 2 external integrations
with the central one.
}
\examples{
path <- system.file("extdata", "ex_annotated_ISMatrix.tsv.xz",
package = "ISAnalytics")
matrix <- import_single_Vispa2Matrix(path)
near <- compute_near_integrations(matrix)
}
\concept{Recalibration functions}
