% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\name{date_formats}
\alias{date_formats}
\title{Possible choices for the \code{dates_format} parameter in
\code{import_association_file},
\code{import_parallel_vispa2Matrices_interactive} and
\code{import_parallel_vispa2Matrices_auto}.}
\usage{
date_formats()
}
\value{
A character vector
}
\description{
All options correspond to \code{lubridate} functions, see more in the dedicated
package documentation.
}
\examples{
date_formats()
}
\seealso{
\code{\link{import_association_file}},
\code{\link{import_parallel_Vispa2Matrices_auto}}

Other Import functions helpers: 
\code{\link{annotation_issues}()},
\code{\link{default_af_transform}()},
\code{\link{default_iss_file_prefixes}()},
\code{\link{matching_options}()},
\code{\link{quantification_types}()}
}
\concept{Import functions helpers}
