% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\name{.auto_detect_type}
\alias{.auto_detect_type}
\title{Internal function to auto-detect the type of IS based on the headers.}
\usage{
.auto_detect_type(df)
}
\arguments{
\item{df}{the data frame to inspect}
}
\value{
one value among:
\itemize{
\item "OLD" : for old-style matrices that had only one column holding all genomic
coordinates
\item "NEW_ANNOTATED" :  for the classic Vispa2 annotated matrices
\item "NEW_NOTANN" : for Vispa2 not annotated matrices
\item "MALFORMED" : in any other case
}
}
\description{
Internal function to auto-detect the type of IS based on the headers.
}
