% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\name{.messy_to_tidy}
\alias{.messy_to_tidy}
\title{Internal function to convert a messy matrix to a tidy data frame}
\usage{
.messy_to_tidy(df)
}
\arguments{
\item{df}{ISADataFrame to convert to tidy}
}
\value{
a tidy ISADataFrame
}
\description{
The function uses the suite of functions provided by the
tidyverse to have a more dense and ordered structure. This function is not
exported and should be called in other importing functions.
}
