% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.messy_to_tidy}
\alias{.messy_to_tidy}
\title{Internal function to convert a messy matrix to a tidy data frame}
\usage{
.messy_to_tidy(df)
}
\arguments{
\item{df}{Messy tibble to convert to tidy}
}
\value{
a tidy tibble
}
\description{
Uses the suite of functions provided by the
tidyverse to produce a more dense and ordered structure.
This function is not exported and should be called in other importing
functions.
}
\keyword{internal}
