% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.read_and_correctness_af}
\alias{.read_and_correctness_af}
\title{Reads association file and checks if it's correct or not.}
\usage{
.read_and_correctness_af(path)
}
\arguments{
\item{path}{The path to the association file on disk}
}
\value{
A tibble containing the association file.
}
\description{
Reads association file and checks if it's correct or not.
}
