% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{generate_blank_association_file}
\alias{generate_blank_association_file}
\title{Creates a blank association file.}
\usage{
generate_blank_association_file(path)
}
\arguments{
\item{path}{The path on disk where the file should be written}
}
\value{
returns NULL
}
\description{
This function is useful if you want a blank association file to start using
both Vispa2 and this package or simply if you want a correct framework to
fix a malformed association file you have already.
}
\examples{
temp <- tempfile()
generate_blank_association_file(temp)
}
\seealso{
Other Utility functions: 
\code{\link{generate_Vispa2_launch_AF}()},
\code{\link{unzip_file_system}()}
}
\concept{Utility functions}
