% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{generate_Vispa2_launch_AF}
\alias{generate_Vispa2_launch_AF}
\title{Creates a reduced association file for Vispa2 run, given project and pool}
\usage{
generate_Vispa2_launch_AF(association_file, project, pool, path)
}
\arguments{
\item{association_file}{The imported association file (via
\code{import_association_file})}

\item{project}{A vector of characters containing project names}

\item{pool}{A vector of characters containing pool names. \strong{NOTE: the names
should refer to the values contained in the PoolID column of the association
file and NOT the concatenatePoolIDSeqRun column!}}

\item{path}{A single string representing the path to the folder where files
should be written. If the folder doesn't exist it will be created.}
}
\value{
returns NULL
}
\description{
The function selects the appropriate columns and prepares a file for the
launch of Vispa2 pipeline for each project/pool pair specified.
}
\details{
Note: the function is vectorized, meaning you can specify more than
one project and more than one pool as vectors of characters, but you must
ensure that:
\itemize{
\item Both \code{project} and \code{pool} vectors have the same length
\item You correclty type names in corresponding positions, for example
c("CLOEXP", "PROJECT1100", "PROJECT1100") - c("POOL6", "ABX-LR-PL5-POOL14-1",
"ABX-LR-PL6-POOL15-1"). If you type a pool in the position of a corresponding
project that doesn't match no file will be produced since that pool doesn't
exist in the corresponding project.
}
}
\examples{
temp <- tempdir()
path_af <- system.file("extdata", "ex_association_file.tsv",
package = "ISAnalytics")
root_pth <- system.file("extdata", "fs.zip", package = "ISAnalytics")
root <- tempdir()
zip::unzip(root_pth, exdir = root)
root <- file.path(root, "fs")
root <- gsub('"', "", gsub("\\\\\\\\", "/", root))
association_file <- import_association_file(path_af, root)
generate_Vispa2_launch_AF(association_file, "CLOEXP", "POOL6", temp)
}
