% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\name{import_association_file}
\alias{import_association_file}
\title{Import the association file from disk}
\usage{
import_association_file(path, root)
}
\arguments{
\item{path}{The path on disk to the association file.}

\item{root}{The path on disk of the root folder of Vispa2 output.
See details.}
}
\value{
A tibble with the contents of the association file plus a column
containing the path in the file system for every project and pool if found.
}
\description{
Imports the association file and immediately performs a check on
the file system starting from the root to assess the alignment between the
two.
}
\details{
The import series of functions is designed to work in combination
with the use of Vispa2 pipeline, please refer to this article for more
details: \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5702242/}{VISPA2:
A Scalable Pipeline for High-Throughput Identification
and Annotation of Vector Integration Sites}.\cr
The pipeline automatically produces an hierarchical structure in the file
system which follows this schema:
\itemize{
\item /root_folder
\itemize{
\item Optional intermediate folders
\itemize{
\item ProjectID\cr
|_bam\cr
|_bcmuxall\cr
|_bed\cr
|_iss\cr
|_quality\cr
|_report\cr
|_quantification\cr
*|___concatenatePoolIDSeqRun\cr
}
}
}

For each ProjectID there may be several nested PoolIDs. The alignment
function only looks for PoolIDs in the quantification folder, since it's
the location of the matrices to import.
For more details on how to properly use these functions, refer to the
vignette - vignette("how_to_import_functions")
}
\examples{
path <- system.file("extdata", "ex_association_file.tsv",
package = "ISAnalytics")
root_pth <- system.file("extdata", "fs.zip", package = "ISAnalytics")
root <- unzip_file_system(root_pth, "fs")
association_file <- import_association_file(path, root)
}
\seealso{
Other Import functions: 
\code{\link{import_parallel_Vispa2Matrices_interactive}()},
\code{\link{import_single_Vispa2Matrix}()}
}
\concept{Import functions}
