% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\name{import_parallel_Vispa2Matrices}
\alias{import_parallel_Vispa2Matrices}
\title{Import integration matrices from paths in the association file.}
\usage{
import_parallel_Vispa2Matrices(
  association_file,
  quantification_type = c("seqCount", "fragmentEstimate"),
  matrix_type = c("annotated", "not_annotated"),
  workers = 2,
  multi_quant_matrix = TRUE,
  report_path = default_report_path(),
  patterns = NULL,
  matching_opt = matching_options(),
  mode = c("AUTO", "INTERACTIVE"),
  ...
)
}
\arguments{
\item{association_file}{Data frame imported via
\link{import_association_file} (with file system alignment) or
a string containing the path to the association file on disk.}

\item{quantification_type}{A vector of requested quantification_types.
Possible choices are \link{quantification_types}}

\item{matrix_type}{A single string representing the type of matrices
to be imported. Can only be one in "annotated" or "not_annotated".}

\item{workers}{A single integer representing the number of parallel
workers to use for the import}

\item{multi_quant_matrix}{If set to TRUE will produce a
multi-quantification matrix through \link{comparison_matrix}
instead of a list.}

\item{report_path}{The path where the report file should be saved.
Can be a folder, a file or NULL if no report should be produced.
Defaults to \code{{user_home}/ISAnalytics_reports}.}

\item{patterns}{Relevant only if argument \code{mode} is set to \code{AUTO}.
A character vector of additional patterns to match on file
names. Please note that patterns must be regular expressions. Can be NULL if
no patterns need to be matched.}

\item{matching_opt}{Relevant only if argument \code{mode} is set to \code{AUTO}.
A single value between \link{matching_options}}

\item{mode}{A single value between \code{AUTO} and \code{INTERACTIVE}. If
INTERACTIVE, the function will ask for input from the user on console,
otherwise the process is fully automated (with limitations, see vignette).}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Additional named arguments
to pass to \code{ìmport_association_file}, \code{comparison_matrix} and
\code{import_single_Vispa2_matrix}}
}
\value{
Either a multi-quantification matrix or a list of integration
matrices
}
\description{
\lifecycle{stable}
The function offers a convenient way of importing multiple integration
matrices in an automated or semi-automated way.
For more details see the "How to use import functions" vignette:
\code{vignette("import_functions_howto", package = "ISAnalytics")}
}
\examples{
fs_path <- system.file("extdata", "fs.zip", package = "ISAnalytics")
fs <- unzip_file_system(fs_path, "fs")
af_path <- system.file("extdata", "asso.file.tsv.gz",
    package = "ISAnalytics"
)
af <- import_association_file(af_path,
    root = fs,
    import_iss = FALSE,
    report_path = NULL
)
matrices <- import_parallel_Vispa2Matrices(af,
    c("seqCount", "fragmentEstimate"),
    mode = "AUTO", report_path = NULL
)
head(matrices)
}
\seealso{
Other Import functions: 
\code{\link{import_Vispa2_stats}()},
\code{\link{import_association_file}()},
\code{\link{import_single_Vispa2Matrix}()}
}
\concept{Import functions}
