% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\name{import_parallel_Vispa2Matrices_auto}
\alias{import_parallel_Vispa2Matrices_auto}
\title{Import integration matrices from association file.}
\usage{
import_parallel_Vispa2Matrices_auto(
  association_file,
  quantification_type,
  matrix_type = "annotated",
  workers = 2,
  multi_quant_matrix = TRUE,
  patterns = NULL,
  matching_opt = matching_options(),
  export_report_path = NULL,
  ...
)
}
\value{
A data frame or a list
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated to avoid redundancy.
Please refer to \code{\link{import_parallel_Vispa2Matrices}}.
}
\keyword{internal}
