% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\name{import_single_Vispa2Matrix}
\alias{import_single_Vispa2Matrix}
\title{Imports a single integration matrix from file.}
\usage{
import_single_Vispa2Matrix(path)
}
\arguments{
\item{path}{the path to the file on disk}
}
\value{
a tidy ISADataFrame
}
\description{
This function allows to read and import an integration matrix
produced as the output of Vispa2 pipeline and converts it to a tidy
ISADataFrame.
}
\examples{
path_to_file <- system.file("extdata", "ex_annotated_ISMatrix.tsv.xz",
package = "ISAnalytics")
isa_dataframe <- import_single_Vispa2Matrix(path_to_file)
}
