% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\name{import_single_Vispa2Matrix}
\alias{import_single_Vispa2Matrix}
\title{Import a single integration matrix from file}
\usage{
import_single_Vispa2Matrix(
  path,
  to_exclude = NULL,
  keep_excluded = FALSE,
  separator = "\\t"
)
}
\arguments{
\item{path}{The path to the file on disk}

\item{to_exclude}{Either \code{NULL} or a character vector of column names that
should be ignored when importing}

\item{keep_excluded}{Keep the columns in \code{to_exclude} as additional
columns?}

\item{separator}{The column delimiter used, defaults to \verb{\\t}}
}
\value{
A data.table object in tidy format
}
\description{
\lifecycle{stable}
This function allows to read and import an integration matrix
(ideally produced by VISPA2) and converts it to a tidy
format.
}
\details{
For more details see the "How to use import functions" vignette:
\code{vignette("import_functions_howto", package = "ISAnalytics")}
}
\section{Required vars and tags}{

\subsection{Required vars}{

TODO
}
}

\examples{
fs_path <- system.file("extdata", "fs.zip", package = "ISAnalytics")
fs <- unzip_file_system(fs_path, "fs")
matrix_path <- fs::path(
    fs,
    "PJ01",
    "quantification",
    "POOL01-1",
    "PJ01_POOL01-1_seqCount_matrix.no0.annotated.tsv.gz"
)
matrix <- import_single_Vispa2Matrix(matrix_path)
head(matrix)
}
\seealso{
Other Import functions: 
\code{\link{import_Vispa2_stats}()},
\code{\link{import_association_file}()},
\code{\link{import_parallel_Vispa2Matrices}()}
}
\concept{Import functions}
