% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\name{import_Vispa2_stats}
\alias{import_Vispa2_stats}
\title{Import Vispa2 stats given the aligned association file.}
\usage{
import_Vispa2_stats(
  association_file,
  file_prefixes = default_iss_file_prefixes(),
  join_with_af = TRUE,
  pool_col = "concatenatePoolIDSeqRun",
  export_widget_path = NULL
)
}
\arguments{
\item{association_file}{The file system aligned association file
(contains columns with absolute paths to the 'iss' folder)}

\item{file_prefixes}{A character vector with known file prefixes
to match on file names. NOTE: the elements represent regular expressions.
For defaults see \link{default_iss_file_prefixes}.}

\item{join_with_af}{Logical, if TRUE the imported stats files will be
merged with the association file, if false a single data frame holding
only the stats will be returned.}

\item{pool_col}{A single string. What is the name of the pool column
used in the Vispa2 run? This will be used as a key to perform a join
operation with the stats files \code{POOL} column.}

\item{export_widget_path}{Either NULL or the path on disk where the
widget report should be saved.}
}
\value{
A data frame
}
\description{
\lifecycle{experimental}
Imports all the Vispa2 stats files for each pool provided the association
file has been aligned with the file system
(see \code{\link{import_association_file}}).
}
\examples{
op <- options("ISAnalytics.widgets" = FALSE)
path <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_pth <- system.file("extdata", "fs.zip", package = "ISAnalytics")
root <- unzip_file_system(root_pth, "fs")
association_file <- import_association_file(path, root, dates_format = "dmy")
af_with_stats <- import_Vispa2_stats(association_file)
options(op)
}
\seealso{
Other Import functions: 
\code{\link{import_association_file}()},
\code{\link{import_parallel_Vispa2Matrices_auto}()},
\code{\link{import_parallel_Vispa2Matrices_interactive}()},
\code{\link{import_single_Vispa2Matrix}()}
}
\concept{Import functions}
