% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISAnalytics-package.R
\docType{package}
\name{ISAnalytics-package}
\alias{ISAnalytics}
\alias{ISAnalytics-package}
\title{ISAnalytics: Analyze gene therapy vector insertion sites data identified from genomics next generation sequencing reads for clonal tracking studies}
\description{
ISAnalytics is an R package developed to analyze gene therapy vector insertion sites data identified from genomics next generation sequencing reads for clonal tracking studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com//calabrialab/isanalytics}
  \item Report bugs at \url{https://github.com//calabrialab/isanalytics}
}

}
\author{
\strong{Maintainer}: Andrea Calabria \email{calabria.andrea@hsr.it}

Authors:
\itemize{
  \item Giulio Spinozzi \email{spinozzi.giulio@hsr.it}
  \item Giulia Pais \email{giuliapais1@gmail.com}
}

}
\keyword{internal}
