% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported-vars.R
\name{mandatory_IS_vars}
\alias{mandatory_IS_vars}
\title{Names of mandatory variables for an integration matrix.}
\usage{
mandatory_IS_vars()
}
\value{
A character vector
}
\description{
Contains the names of the columns that need to be present in order for a
tibble to be considered an integration matrix.
}
\examples{
mandatory_IS_vars()
}
