% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-reads-filtering.R
\name{outlier_filter}
\alias{outlier_filter}
\title{Filter out outliers in metadata, identified by the chosen outlier test.}
\usage{
outlier_filter(
  metadata,
  outlier_test = "outliers_by_pool_fragments",
  negate = FALSE,
  ...
)
}
\arguments{
\item{metadata}{The metadata data frame}

\item{outlier_test}{A string representing a function name. The name
must be one of the available outlier tests, see
\link{available_outlier_tests}.}

\item{negate}{If TRUE will return only the metadata that was flagged to
be removed. If FALSE will return only the metadata that wasn't flagged
to be removed.}

\item{...}{Additional named arguments passed to \code{outliers_test}}
}
\value{
A data frame of metadata which has less or the same amount of rows
}
\description{
\lifecycle{experimental}
Filter out outliers in metadata.
}
\examples{
data("association_file", package = "ISAnalytics")
filtered_af <- outlier_filter(association_file,
  key = "BARCODE_MUX",
  report_path = NULL
)
head(filtered_af)
}
\concept{Outliers filter}
