% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{print.ISADataFrame}
\alias{print.ISADataFrame}
\title{Printing ISADataFrames}
\usage{
\method{print}{ISADataFrame}(x, ...)
}
\arguments{
\item{x}{an ISADataFrame object}

\item{...}{optional arguments to print}
}
\value{
Nothing
}
\description{
Printing ISADataFrames
}
\examples{
expList <- list(
    chr = c(as.character(1:10)),
    integration_locus = runif(10, min = 100, max = 10000),
    strand = sample(c("+", "-"), 10, replace = TRUE),
    meta1 = rep_len("m1", 10),
    exp_1 = runif(10, min = 0, max = 10000),
    exp_2 = runif(10, min = 0, max = 10000),
    exp_3 = runif(10, min = 0, max = 10000)
)
isadf <- ISADataFrame(expList, metadata = c("meta1"))
print(isadf)
}
