% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collision-removal.R
\name{realign_after_collisions}
\alias{realign_after_collisions}
\title{Re-aligns matrices of other quantification types based on the processed
sequence count matrix.}
\usage{
realign_after_collisions(sc_matrix, other_matrices)
}
\arguments{
\item{sc_matrix}{The sequence count matrix already processed for collisions
via \code{remove_collisions}}

\item{other_matrices}{A named list of matrices to re-align. Names in the list
must be quantification types (\code{quantification_types()}) except
"seqCount".}
}
\value{
A named list with re-aligned matrices
}
\description{
\lifecycle{experimental}
This function should be used to keep data consistent among the same analysis:
if for some reason you removed the collisions by passing only the sequence
count matrix to the \code{remove_collisions} function, you should call this
function afterwards, providing a list of other quantification matrices.
NOTE: if you provided a list of several quantification types to
\code{remove_collisions} before, there is no need to call this function.
}
\details{
For more details on how to use collision removal functionality:
\code{vignette("Collision removal functionality", package = "ISAnalytics")}
}
\examples{
op <- options("ISAnalytics.widgets" = FALSE)
path <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_pth <- system.file("extdata", "fs.zip", package = "ISAnalytics")
root <- unzip_file_system(root_pth, "fs")
association_file <- import_association_file(path, root)
matrices <- import_parallel_Vispa2Matrices_auto(
    association_file, NULL,
    c("fragmentEstimate", "seqCount"), "annotated", 2, NULL, "ANY"
)
sc_matrix <- remove_collisions(matrices$seqCount, association_file)
others <- matrices[!names(matrices) \%in\% "seqCount"]
aligned_matrices <- realign_after_collisions(sc_matrix, others)
options(op)
}
\seealso{
\code{\link{remove_collisions}}

Other Collision removal: 
\code{\link{remove_collisions}()}
}
\concept{Collision removal}
