% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collision-removal.R
\name{remove_collisions}
\alias{remove_collisions}
\title{Identifies and removes collisions based on the sequence count matrix.}
\usage{
remove_collisions(
  x,
  association_file,
  date_col = "SequencingDate",
  reads_ratio = 10,
  seq_count_col = "seqCount",
  max_rows_reports = 50
)
}
\arguments{
\item{x}{A named list of matrices (names must be quantification types),
a single integration matrix representing the sequence count matrix of
interest or a multi-quantification matrix obtained via
\link{comparison_matrix}}

\item{association_file}{The association file imported via
\code{import_association_file}}

\item{date_col}{The date column that should be considered for the analysis.
Must be one value in \code{date_columns_coll()}}

\item{reads_ratio}{A single numeric value that represents the ratio that has
to be considered when deciding between seqCount value.}

\item{seq_count_col}{For support of multi-quantification matrix -
the name of the sequence count values column}

\item{max_rows_reports}{A numeric value, represents the maximum number of
rows of the reports data frames that can be printed on console
if the option \code{ISAnalytics.verbose} is active. If the data frames are too
large they won't be printed on console - we recommend using widgets for
detailed and more accessible info.}
}
\value{
A list of tibbles with removed collisions
}
\description{
\lifecycle{experimental}
A collision is an integration (aka a unique combination of
chr, integration_locus and strand) which is observed in more than one
independent sample (a unique pair of ProjectID and SubjectID). The function
tries to decide to which subject an integration should be assigned and if no
decision can be taken, the integration is completely removed from the data
frame.
}
\details{
If you don't want the function to show details and messages do:
\code{options(ISAnalitics.verbose = FALSE)}.
To restore to the original value:
\code{options(ISAnalitics.verbose = TRUE)}.
For more details on how to use collision removal functionality:
\code{vignette("Collision removal functionality", package = "ISAnalytics")}
}
\examples{
op <- options("ISAnalytics.widgets" = FALSE)
path <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_pth <- system.file("extdata", "fs.zip", package = "ISAnalytics")
root <- unzip_file_system(root_pth, "fs")
association_file <- import_association_file(path, root,
    dates_format = "dmy")
matrices <- import_parallel_Vispa2Matrices_auto(
    association_file, NULL,
    c("fragmentEstimate", "seqCount"), "annotated", 2, NULL, "ANY"
)
matrices <- remove_collisions(matrices, association_file)
options(op)
}
\seealso{
\code{\link{date_columns_coll}}

Other Collision removal: 
\code{\link{realign_after_collisions}()}
}
\concept{Collision removal}
