% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collision-removal.R
\name{remove_collisions}
\alias{remove_collisions}
\title{Identifies and removes collisions.}
\usage{
remove_collisions(
  x,
  association_file,
  independent_sample_id = c("ProjectID", "SubjectID"),
  date_col = "SequencingDate",
  reads_ratio = 10,
  quant_cols = c(seqCount = "seqCount", fragmentEstimate = "fragmentEstimate"),
  report_path = default_report_path(),
  max_workers = NULL
)
}
\arguments{
\item{x}{Either a multi-quantification matrix or a
named list of matrices (names must be quantification types)}

\item{association_file}{The association file imported via
\code{import_association_file()}}

\item{date_col}{The date column that should be considered.
Must be one value in \code{date_columns_coll()}}

\item{reads_ratio}{A single numeric value that represents the ratio that has
to be considered when deciding between \code{seqCount} value.}

\item{quant_cols}{A named character vector where names are
quantification types and
values are the names of the corresponding columns. The quantification
\code{seqCount} MUST be included in the vector.}

\item{report_path}{The path where the report file should be saved.
Can be a folder, a file or NULL if no report should be produced.
Defaults to \code{{user_home}/ISAnalytics_reports}.}

\item{max_workers}{Maximum number of parallel workers to distribute the
workload. If \code{NULL} (default) produces the maximum amount of workers allowed,
a numeric value is requested otherwise. WARNING: a higher number of workers
speeds up computation at the cost of memory consumption! Tune this parameter
accordingly.}
}
\value{
Either a multi-quantification matrix or a list of data frames
}
\description{
\lifecycle{stable}
A collision is an integration (aka a unique combination of
\code{chr}, \code{integration_locus} and \code{strand}) which is observed in more than one
independent sample (a unique pair of \code{ProjectID} and \code{SubjectID}).
The function tries to decide to which subject an integration
should be assigned to and, if no
decision can be taken, the integration is completely removed from the data
frame.
For more details refer to the vignette "Collision removal functionality":
\code{vignette("collision_removal", package = "ISAnalytics")}
}
\examples{
data("integration_matrices", package = "ISAnalytics")
data("association_file", package = "ISAnalytics")
no_coll <- remove_collisions(
    x = integration_matrices,
    association_file = association_file,
    report_path = NULL
)
head(no_coll)
}
\seealso{
\code{\link{date_columns_coll}}

Other Collision removal: 
\code{\link{realign_after_collisions}()}
}
\concept{Collision removal}
