% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-functions.R
\name{separate_quant_matrices}
\alias{separate_quant_matrices}
\title{Separate a multiple-quantification matrix into single quantification
matrices.}
\usage{
separate_quant_matrices(
  x,
  fragmentEstimate = "fragmentEstimate",
  seqCount = "seqCount",
  barcodeCount = "barcodeCount",
  cellCount = "cellCount",
  ShsCount = "ShsCount"
)
}
\arguments{
\item{x}{Single integration matrix with multiple quantification
value columns, likely obtained via \link{comparison_matrix}.}

\item{fragmentEstimate}{Name of the fragment estimate values column
in input}

\item{seqCount}{Name of the sequence count values column
in input}

\item{barcodeCount}{Name of the barcode count values column
in input}

\item{cellCount}{Name of the cell count values column
in input}

\item{ShsCount}{Name of the shs count values column
in input}
}
\value{
A named list of tibbles, where names are quantification types
}
\description{
\lifecycle{maturing}
The function separates a single multi-quantification integration
matrix, obtained via \link{comparison_matrix}, into single
quantification matrices as a named list of tibbles.
}
\examples{
op <- options("ISAnalytics.widgets" = FALSE)
path <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_pth <- system.file("extdata", "fs.zip", package = "ISAnalytics")
root <- unzip_file_system(root_pth, "fs")
matrices <- import_parallel_Vispa2Matrices_auto(
    path, root,
    c("fragmentEstimate", "seqCount"), "annotated", 2, NULL, "ANY"
)
total_matrix <- comparison_matrix(matrices)
separated_matrix <- separate_quant_matrices(total_matrix)
options(op)
}
\seealso{
\link{quantification_types}

Other Analysis functions: 
\code{\link{comparison_matrix}()},
\code{\link{compute_abundance}()},
\code{\link{threshold_filter}()},
\code{\link{top_integrations}()}
}
\concept{Analysis functions}
