% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{unzip_file_system}
\alias{unzip_file_system}
\title{A utility function to unzip and use example file systems included in the
package}
\usage{
unzip_file_system(zipfile, name)
}
\arguments{
\item{zipfile}{The zipped file to decompress}

\item{name}{The name of the folder in the zipped archive ("fs" or "fserr")}
}
\value{
A path to reference
}
\description{
This utility function is a simple shortcut to create a temporary directory,
unzip and reference the examples file systems included in the package for
testing purposes.
}
\examples{
root_pth <- system.file("extdata", "fs.zip", package = "ISAnalytics")
root <- unzip_file_system(root_pth, "fs")
}
\seealso{
Other Utility functions: 
\code{\link{as_sparse_matrix}()},
\code{\link{generate_Vispa2_launch_AF}()},
\code{\link{generate_blank_association_file}()}
}
\concept{Utility functions}
