% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{vec_restore.ISADataFrame}
\alias{vec_restore.ISADataFrame}
\title{Implementation of vec_restore for ISADataFrame.}
\usage{
vec_restore.ISADataFrame(x, to, ...)
}
\arguments{
\item{x}{A vector.}

\item{to}{The original vector to restore to.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
See official documentation at \code{
\link[vctrs:vec_proxy]{vec_restore}}
}
\description{
Implementation of vec_restore for ISADataFrame.
}
